% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{AR}
\alias{AR}
\title{Liczy Accuracy Ratio (GINI)}
\usage{
AR(score, def, plot = FALSE, sort.order = -1, label = "", ...)
}
\arguments{
\item{score}{wektor    wartoœci, wed³ug który nale¿y posortowaæ wartoœci \code{def}.}

\item{def}{Wektor wartoœci \code{\{0,1\}}.}

\item{plot}{Czy narysowaæ krzyw¹ CAP. Domyœlnie \code{FALSE}.}

\item{sort.order}{Dla \code{sort.order=1} sortuje rosn¹co, a dla \code{sort.order=-1} malej¹co
po \code{score}.}

\item{label}{opis wyœwietlony w legendzie.}

\item{...}{dodatkowe parametry.}
}
\value{
Zwraca wektor ze statystykami jakoœci modelu.
 \item{AUCAP}{Powierzchnia pod krzyw¹ CAP.}
 \item{AR}{Wyliczone Accuracy Ratio (GINI).}
 \item{AUC}{Powierzchnia pod krzyw¹ ROC.}							
 \item{KS}{Statystyka Ko³mogorowa-Smirnowa.}
}
\description{
Wylicza Accuracy Ratio (GINI). Funkcja jest znacznie uproszczona w porównaniu z poprzedni¹ wersj¹ 
i funkcje nie s¹ ze sob¹ kompatybilne. Najwiêksz¹ zmian¹ jest wykorzystanie biblioteki \code{data.table},
dziêki czemu nieporównywalnie zwiêkszona zosta³a prêdkoœæ dzia³ania.
}
\author{
Micha³ Danaj
}

