% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{AR_quick}
\alias{AR_quick}
\title{Liczy AR(GINI) dla specyficznie przygotowanych danych}
\usage{
AR_quick(bad, obs)
}
\arguments{
\item{bad}{liczba badów odpowiadaj¹ca kolejnym wartoœciom score.}

\item{obs}{liczba obserwacji odpowiadaj¹ca kolejnym wartoœciom score.}
}
\description{
Funkcja zak³ada, ¿e podane licznoœci badów i wszystich obserwacji
posortowane s¹ wg rosn¹cych wartoœci score. W przypadku gdy w próbie s¹
tylko bady lub tylko goody, funkcja zwraca \code{NaN}.
}
\author{
Micha³ Danaj
}

