% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{as.data.frame.AR}
\alias{as.data.frame.AR}
\title{Przekszta³ca obiekt na data.frame}
\usage{
\method{as.data.frame}{AR}(x)
}
\arguments{
\item{x}{obiekt klasy \code{AR}.}
}
\description{
Przekszta³ca obiekt na \code{\link{data.frame}}
}
\details{
Przekszta³ca obiekt na \code{\link{data.frame}}
}

