% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreCard.r
\name{assignScore}
\alias{assignScore}
\title{Przypisuje score z definicji karty scoringowej.}
\usage{
assignScore(scoreCard, x, sufix = "_points")
}
\arguments{
\item{scoreCard}{defini`cja karty scoringowej. Patrz \code{\link{getScoreCard}}.}

\item{x}{\code{data.frame} z kolumnami o nazwach takich, jak w definicji karty}

\item{sufix}{ciąg znaków dodany do nazw wynikowych kolumn.}
}
\description{
Przypisuje score z definicji karty scoringowej.
}
\seealso{
\code{\link{getScoreCard}}.
}
