% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make.r
\name{bckt_br}
\alias{bckt_br}
\title{Dzieli na przedziały jedną z dwóch metod i wylicza na nich bad rate (dev).}
\usage{
bckt_br(x, y, n, weights = rep(1, length(x)), method = c("eq_length",
  "eq_count"), one.value.action = c("none", "combine"), avg = NULL,
  total = TRUE, sort_x = TRUE)
}
\arguments{
\item{x}{Zmienna, którą będziemy dyskretyzować.}

\item{y}{Zmienna dwumianowa.}

\item{n}{Liczba wynikowych przedziałów.}

\item{weights}{wektor z wagami.}

\item{method}{Sposób podziału na przedziały. Szczegóły w sekcji Details.}

\item{one.value.action}{Jeszcze nie działa.}

\item{avg}{\code{vector}, \code{data.frame}, \code{data.table} z wartościami dla których
zostaną wyliczone średnie.}

\item{total}{czy dołączyć wiersz z podsumowaniem.}

\item{sort_x}{jeśli \code{TRUE}, wynikowa tabela zostanie posortowana po wartościach}
}
\value{
Zwraca \code{data.frame}, w którym dla \code{i}-tego wiersza podane są statystyki
dla \code{i}-tego przedziału.
}
\description{
Wersja w developmencie.
Może poza tym, że jeśli nie ma obserwacji w jakimś bukcecie, to go usuwa.
}
\author{
Michał Danaj
}
