% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make.r
\name{bckt_stat}
\alias{bckt_stat}
\title{Wylicza statystyki zmiennej \code{y} dla podanych grup \code{x} (dev)}
\usage{
bckt_stat(x = NULL, y = NULL, weights = rep(1, length(x)), dt = NULL,
  avg = NULL, total = TRUE, sort_x = TRUE)
}
\arguments{
\item{x}{wektor z identyfikatorami bucketów.Nie może być braków danych.}

\item{y}{wektor zmiennych numerycznych. Z założenia zmienna dwumianowa, ale statystyki
zostaną policzone dla każdej innej zmiennej numerycznej.Nie może być braków danych.}

\item{weights}{wektor z wagami.}

\item{dt}{\code{data.table} z kolumnami \code{x}, \code{y}, \code{weights}.}

\item{avg}{\code{vector}, \code{data.frame}, \code{data.table} z wartościami dla których
zostaną wyliczone średnie.}

\item{total}{czy dołączyć wiersz z podsumowaniem.}

\item{sort_x}{jeśli \code{TRUE}, wynikowa tabela zostanie posortowana po wartościach 
cechy \code{x}. W przeciwnym wypadku sortowanie będzie zgodne z porządkiem w danych.}
}
\value{
Zwraca \code{data.table} ze statystykami.
}
\description{
Wylicza statystyki zmiennej \code{y} dla podanych grup \code{x}. W założeniu
\code{y} jest zmienną dwumianową o wartościach {0,1}.
}
\author{
Michał Danaj
}
