% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make.r
\name{bckt_stat2}
\alias{bckt_stat2}
\title{Wylicza statystyki dla podanych przedziałów (dev)}
\usage{
bckt_stat2(breaks, x = NULL, y = NULL, weights = rep(1, length(x)),
  dt = NULL, avg = NULL, total = TRUE, sort_x = TRUE)
}
\arguments{
\item{breaks}{punkty podziału}

\item{x}{zmienna score'owa.}

\item{y}{zmienna odpowiedzi z zakresu [0,1]. Np. default, LGD.}

\item{weights}{wektor z wagami.}

\item{dt}{\code{data.table} z kolumnami \code{x}, \code{y}, \code{weights}.}

\item{avg}{\code{vector}, \code{data.frame}, \code{data.table} z wartościami dla których
zostaną wyliczone średnie.}

\item{total}{czy dołączyć wiersz z podsumowaniem.}

\item{sort_x}{jeśli \code{TRUE}, wynikowa tabela zostanie posortowana po wartościach}
}
\value{
\code{data.frame} ze statystykami.
}
\description{
Dzieli zmienną \code{x} na przedziały określone przez \code{breaks}
i wylicza statystyki dla tak powstałych bucketów.
}
\details{
Do przypisania wartości do przedziału używana jest funkcja

\code{findInterval(x, y, rightmost.closed = TRUE, all.inside = TRUE)}.
}
\seealso{
\code{\link{buckety_stat}}.
}
\author{
Michał Danaj
}
