% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make.R
\name{buckety_stat}
\alias{buckety_stat}
\title{Wylicza statystyki zmiennej \code{default} dla podanych grup \code{buckets}}
\usage{
buckety_stat(bucket, default, total=TRUE)
}
\arguments{
\item{bucket}{wektor z identyfikatorami bucket?w.}

\item{default}{wektor zmiennych numerycznych.}

\item{total}{czy do??czy? wiersz z podsumowaniem.}
}
\value{
Zwraca \code{data.frame}, w kt?rym dla \code{i}-tego wiersza podane s? statystyki
  dla \code{i}-tego przedzia?u.
  \item{nr }{Numer przedzia?u.}
  \item{od }{Pocz?tek przedzia?u.}
  \item{srodek}{?rodek przedzia?u.}
  \item{mean}{?rednia z obserwacji w przedziale.}
  \item{do}{Koniec przedzia?u.}
  \item{n_default}{Liczba default?w.}
  \item{n_obs}{Liczba obserwacji.}
  \item{br}{Bad rate.}
  \item{logit}{Logit.}
  \item{probit}{Probit.}
  \item{var}{Wariancja zmiennej \code{default}.}
  \item{waga}{Waga jako odwrotno?? warinacji. Gdy \code{n_default=0} to przyjmuje si?,
			?e \code{n_default=0.5}}
}
\description{
Wylicza statystyki zmiennej \code{default} dla podanych grup \code{buckets}
}
\author{
Micha? Danaj
}
