% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make_tree.R
\name{discretization}
\alias{discretization}
\title{Dyskretyzuje zmienn¹ ciag³¹ drzewkiem w oparciu o zmienn¹ odpowiedzi}
\usage{
discretization(x, y, special_val = NULL, min_bucket = 20, max_gleb = 3,
  interactive = FALSE, breaks = NULL, testy = FALSE, ...)
}
\arguments{
\item{x}{zmienna ci¹g³a.}

\item{y}{zmienna celu (PD, LGD).}

\item{special_val}{wartoœci specjalne, które powinny zostaæ uwzglêdnione jako
osobne klasy. Wartoœæ tak¹ jest równie¿ \code{NA}, automatycznie
uwzglêdniana jako osobna klasa.}

\item{min_bucket}{Minimalna wielkoœæ liœcia}

\item{max_gleb}{Maksymalna g³êbokoœc do której budujemy drzewo}

\item{interactive}{TRUE, jeœli zmienna ma byæ dyskretyzowana interaktywnie. W
przeciwnym razie, co jest wartoœci¹ domyœln¹, dyskretyzacja
jest automatyczna.}

\item{breaks}{Zamiast automatycznego dzielenia, mo¿na podaæ wartoœci przedzia³ów (from,to].}

\item{testy}{Wyœwietla komunikaty testowe.}

\item{...}{inne parametry do funkcji \code{\link{drzewo}}.}
}
\description{
W parametrze \code{x} ani \code{y} nie mo¿e byæ NULLi. Mo¿na je jakoœ zakodowaæ.
}
\author{
Micha³ Danaj
}
\seealso{
\code{drzewo}
}

