% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{dopasowanie_do_zmiennej}
\alias{dopasowanie_do_zmiennej}
\title{Sprawdzenie, czy estymacja modelu dobrze t³umaczy zale¿noœæ targetu od zmiennej}
\usage{
dopasowanie_do_zmiennej(x, y, bucket, subset = NULL, ...)
}
\arguments{
\item{x}{wektror z wartoœciami zmiennej}

\item{y}{wektor z predykcj¹ modelu}

\item{bucket}{grupowanie zmiennej}

\item{subset}{podzbiór}

\item{...}{dodatkowe parametry graficzne}
}
\value{
bucket z dodanymi wyestymowanymi wartoœciami
}
\description{
Sprawdzenie, czy estymacja modelu dobrze t³umaczy zale¿noœæ targetu od zmiennej
}
\author{
Micha³ Danaj
}

