% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make_tree.r
\name{drzewo}
\alias{drzewo}
\title{Automatyczna dyskretyzacja w oparciu o algorytm drzewa}
\usage{
drzewo(score, def, freq = NULL, wytnij = 0, min_split = 30,
  min_bucket = 10, max_gleb = 4, n_buckets = 20, plot = TRUE,
  testy = FALSE, ...)
}
\arguments{
\item{score}{Zmienna do dyskretyzacji}

\item{def}{Zmienna celu}

\item{freq}{Liczność. Na razie nie obsługiwane}

\item{wytnij}{Jaką część końcowych wartości usunąć funkcją \code{usun_konce}}

\item{min_split}{Minimalna wielkość węzła, aby móc dokonać podziału}

\item{min_bucket}{Minimalna wielkość liścia}

\item{max_gleb}{Maksymalna głębokośc do której budujemy drzewo}

\item{n_buckets}{Chyba nie używane?}

\item{plot}{Rysuje wynik dyskretyzacji}

\item{testy}{Czy wyświetlać komentarze do testowania}

\item{...}{Dodatkowe parametry do rysowania}
}
\value{

}
\description{
Dyskretyzuje zmienną \code{score}
}
\author{
Michał Danaj
}
