% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make_tree.R
\name{drzewo_podzial}
\alias{drzewo_podzial}
\title{Podzia³ drzewa}
\usage{
drzewo_podzial(score, def, nr_wezla, od, do, freq, glebokosc, min_split = 200,
  min_bucket = 100, max_gleb = 3, testy = FALSE)
}
\arguments{
\item{score}{}

\item{def}{}

\item{nr_wezla}{}

\item{od}{}

\item{do}{}

\item{freq}{}

\item{glebokosc}{}

\item{min_split}{}

\item{min_bucket}{}

\item{max_gleb}{}

\item{testy}{}
}
\value{

}
\description{
Podzia³ drzewa
}
\author{
Piotr
}

