% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raport.R
\name{genRaport}
\alias{genRaport}
\title{Generuje raport}
\usage{
genRaport(wyniki, dir = tempdir(), kolejnosc = 1:length(wyniki),
  show = TRUE, scale = c(0, 0.2))
}
\arguments{
\item{wyniki}{lista z wynikami dyskretyzacji itp, z funkcji \code{\link{univariate_anal_stats}}}

\item{dir}{katalog z raportem, jako pe³na bezwzglêdna œcie¿ka! Katalog musi byæ stworzony. Domyœlnie \code{\link{tempdir()}}.}

\item{kolejnosc}{kolejnoœæ wg której zmienne maj¹ byæ wyœwietlone.}

\item{show}{czy na koniec wyœwietliæ raport w przegl¹darce. Domyœlnie \code{TRUE}.}

\item{scale}{Skala osi OY.}
}
\description{
Generuje raport
}
\author{
Micha³ Danaj
}

