% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_SQL.R
\name{genSQL}
\alias{genSQL}
\title{Na podstawie dyskretyzacji i modelu generuje SQL do implementacji}
\usage{
genSQL(wyniki, dane, model)
}
\arguments{
\item{wyniki}{wyniki dyskretyzacji.}

\item{dane}{dane ze zmiennymi z modelu do określenia typu danych.}

\item{model}{model logistyczny.}
}
\value{
Zwraca gotowego sql-a oraz kopiuje go do schowka.
}
\description{
Na podstawie dyskretyzacji i modelu generuje SQL do implementacji
}
\author{
Michał Danaj
}
