% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_SQL.R
\name{genSQLCase}
\alias{genSQLCase}
\title{Generuje kod SQL dla dyskretyzacji jednej zmiennej.}
\usage{
genSQLCase(bucket, nazwaZmiennej, typZmiennej, nazwaZmiennejOut)
}
\arguments{
\item{bucket}{dyskretyzacja zmiennej.}

\item{nazwaZmiennej}{nazwa zmiennej w danych do score'owania.}

\item{typZmiennej}{jeœli zmienna jest znakowa, powinna przyj¹æ wartoœæ \code{'character'}.}

\item{nazwaZmiennejOut}{nazwa wynikowej zmiennej.}
}
\value{
zwraca \code{vector} znakowy z kodem SQL.
}
\description{
Generuje kod SQL dla dyskretyzacji jednej zmiennej na podstawie danych w \code{bucket},
przypisuj¹cy woe.
}
\author{
Micha³ Danaj
}

