% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_SQL.r
\name{genSQLCases}
\alias{genSQLCases}
\title{Generuję listę kodów SQL z dyskretyzacją wszystkich zmiennych z \code{wynik}..}
\usage{
genSQLCases(wynik, dane, varNames)
}
\arguments{
\item{dane}{dane ze zmiennymi z modelu do określenia typu danych.}

\item{varNames}{\code{vector} z nazwami zmiennych wynikowych.}

\item{wyniki}{wyniki dyskretyzacji.}
}
\value{
zwraca listę z kodami dla poszczególnych zmiennych.
}
\description{
Generuję listę kodów SQL z dyskretyzacją wszystkich zmiennych z \code{wynik}.
}
\author{
Michał Danaj
}
