% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_SQL.R
\name{genSQLCases2}
\alias{genSQLCases2}
\title{Generuję listę kodów SQL z dyskretyzacją wszystkich zmiennych z \code{wynik}..}
\usage{
genSQLCases2(wynik, dane, mapping)
}
\arguments{
\item{dane}{dane ze zmiennymi z modelu do określenia typu danych.}

\item{mapping}{mapowanie nazwy źródłowej (przed zastosowniem dyskretyzacji z \code{wynik}
    na nazwę, która jest później w modelu. Jeśli jedna i druga nazwa jest taka sama,
to znaczy że nie było robionej dyskretyzacji i case jest dla tej zmeinnej
pominięty.}

\item{wyniki}{wyniki dyskretyzacji.}
}
\value{
zwraca listę z kodami dla poszczególnych zmiennych.
}
\description{
Generuję listę kodów SQL z dyskretyzacją wszystkich zmiennych z \code{wynik}.
}
\author{
Michał Danaj
}
