% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreCard.R
\name{getScoreCard}
\alias{getScoreCard}
\title{Wylicza punkty score'owe na podstawie parametrów modelu}
\usage{
getScoreCard(model, from, to, test = FALSE)
}
\arguments{
\item{model}{model, którego parametry zostan¹ przekszta³cone na punkty score'owe.}

\item{from}{od jakiej wartoœci maj¹ siê rozpoczynaæ punkty score'owe.}

\item{to}{do jakiej wielkoœci maj¹ byæ wartoœci punktów socre'owych.}

\item{test}{jeœli TRUE, zwraca testy istotnoœci atrybutów.}
}
\description{
Wylicza punkty score'owe na podstawie parametrów modelu
}

