% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreCard.r
\name{getScoreCard2}
\alias{getScoreCard2}
\title{Wylicza punkty score'owe na podstawie parametrów modelu}
\usage{
getScoreCard2(model, from, to, test = FALSE)
}
\arguments{
\item{model}{model, którego parametry zostaną przekształcone na punkty score'owe.}

\item{from}{od jakiej wartości mają się rozpoczynać punkty score'owe.}

\item{to}{do jakiej wielkości mają być wartości punktów socre'owych.}

\item{test}{jeśli TRUE, zwraca testy istotności atrybutów.}
}
\description{
Wylicza punkty score'owe na podstawie parametrów modelu
}
