% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{getStats.AR}
\alias{getStats.AR}
\title{Zwraca statystyki przechowywane w obiekcie AR}
\usage{
getStats.AR(x)
}
\arguments{
\item{x}{obiekt klasy \code{AR}.}
}
\description{
Zwraca statystyki przechowywane w obiekcie \code{\link{AR}}
}
\details{
Zwraca statystyki przechowywane w obiekcie \code{\link{AR}}
}
