% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{HTML.AR}
\alias{HTML.AR}
\title{Wyświetla statystyki przechowywane w obiekcie \code{\link{AR}}}
\usage{
\method{HTML}{AR}(x)
}
\arguments{
\item{x}{obiekt klasy \code{AR}.}
}
\description{
Wyświetla statystyki przechowywane w obiekcie \code{\link{AR}}
}
