% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{informacje}
\alias{informacje}
\title{Rysuje informacje o danych}
\usage{
informacje(
  score,
  default,
  buckets = 20,
  span = 0.8,
  main = "",
  hist_col = "blue",
  ylab = c("Frequency", "BR"),
  xlab = "Calibrated score",
  ...
)
}
\arguments{
\item{score}{Wektor ze zmienną numeryczną.}

\item{default}{Wektor ze zmienną dwumianową.}

\item{buckets}{Sugerowana liczba bucketów.}

\item{span}{Współczynnik wygładzający, wykorzystywany przez funkcję \code{\link[locfit]{locfit}}}

\item{main}{Tytuł wykresu.}

\item{hist_col}{Kolor histogramu}

\item{ylab}{Label}

\item{xlab}{Label}

\item{...}{dodatkowe parametry.}
}
\description{
Rysuje histogram zmiennej \code{score}, dla każdego przedziału histogramu
 rysuje bad rate, oraz rysuję krzywą regresji dla \code{default~score}. Pod
 wykresem wyświetla informacje o Accuracy Ratios i bad rate.
}
\examples{
x<-rnorm(100);
y<-(x+rnorm(100))>0;
informacje(x, y, buckets=15, span=0.8, main="", hist_col="blue", type="br", method_bucket="eq_length")
}
\author{
Michał Danaj
}
