% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interactionsTestAll}
\alias{interactionsTestAll}
\title{Testuje interakcje pomiêdzy wszystkimi zmiennymi}
\usage{
interactionsTestAll(data, discr_list, target = data$target,
  vars_names = names(discr_list), colname = "label")
}
\arguments{
\item{data}{\code{data.frame} z danymi.}

\item{discr_list}{lista z opisem dyskretyzacji.}

\item{target}{wektor z wartoœciami zmiennej celu}

\item{vars_names}{wektor nazw zmiennych do sprawdzenia interakcji. Domyœlnie wszystkie zmienne z \code{discr_list}.}

\item{colname}{nazwa kolumny z tabeli z definicj¹ dyskretyzacji z której zostan¹ pobrane wartoœci do
przypisania dyskretyzacji zmiennej.}
}
\value{
Zwrócona zostanie lista z dwoma elementami:
\item{pvalue_table}{Zawiera \code{data.frame} z nazw¹ pierwszej zmiennej, drugiej zmiennej, oraz p-value interakcji.}
\item{details}{Zawiera listê ze statystykami zwróconymi przez \code{\link{interactionsTest}}.}
}
\description{
Na podstawie zmiennych Ÿród³owych oraz definicji
dyskretyzacji z \code{discr_list}, funkcja przypisuje zmiennym ich wartoœci dyskretne oraz
testuje interakcjê ka¿dej pary ze zmiennych okreœlonych w \code{vars_names}. Do testowania korzysta z funkcji \code{\link{interactionsTest}}.
}
\author{
Micha³ Danaj
}

