% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interakcja}
\alias{interakcja}
\title{Rysuje interakcje dwóch zmiennych}
\usage{
interakcja(zm1, zm2, target)
}
\arguments{
\item{zm1}{Pierwsza zmienna.}

\item{zm2}{Druga zmienna.}

\item{target}{Zmienna celu.}
}
\value{

}
\description{
Rysuje interakcje dwóch zmiennych
}
\author{
Micha³ Danaj
}

