% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{korelacje_zmiennych}
\alias{korelacje_zmiennych}
\title{Wyœwietla w gridzie wartoœci korelacji zmiennych z modelu}
\usage{
korelacje_zmiennych(model, data)
}
\arguments{
\item{model}{dopasowany model.}

\item{data}{dane.}
}
\value{

}
\description{
Wyœwietla w gridzie wartoœci korelacji zmiennych z modelu
}
\author{
Micha³ Dnaaj
}

