% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{make_model_formula}
\alias{make_model_formula}
\title{Na podstawie nazwy zmiennej celu i zmiennych objaśniających, tworzy formułę modelu}
\usage{
make_model_formula(target, vars)
}
\arguments{
\item{target}{zmienna celu}

\item{vars}{zmienne objaśniające}
}
\value{
forumła do zastosowania w modelu
}
\description{
Na podstawie nazwy zmiennej celu i zmiennych objaśniających, tworzy formułę modelu
}
\author{
Michał Danaj
}
