% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{makeCoarseClassingTables}
\alias{makeCoarseClassingTables}
\title{Generuje tabelê z coarse classing do dokumentacji}
\usage{
makeCoarseClassingTables(wyniki)
}
\arguments{
\item{wyniki}{lista z wynikami dyskretyzacji}
}
\description{
Jeœli Woe lub IV wychoch¹ +-Inf, to wstawia zamiast tego NA
}

