% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_assign.R
\name{mapuj}
\alias{mapuj}
\title{Na podstawie zadanych warunków przypisuje etykiety}
\usage{
mapuj(data, mapping)
}
\arguments{
\item{data}{\code{data.frame} do którego będzie przypisywane mapowanie}

\item{mapping}{\code{data.frame} z dwoma kolumnami znakowymi (lub do przerobienia przez \code{as.character})
\code{war} oraz \code{label}}
}
\description{
Na podstawie zadanych warunków w \code{mapping$war} przypisuje etykietę \code{mapping$label}.
W \code{mapping$label} nie może być wartości pustych stringów. Pusty string stosowany jest jako
brak danych. Jeśli dla jakieś elementu żaden warunek nie zachodzi, zwracany jest pusty string.
}
\details{
Możliwe jest określenie warunku \code{else} poprzez wpisanie w \code{mapping$war} stringa "else".
}
\author{
Michał Danaj
}
