% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\docType{package}
\name{MDBinom-package}
\alias{MDBinom}
\alias{MDBinom-package}
\title{Ró¿ne dzia³ania na zmiennych dwumianowych}
\description{
Ró¿ne dzia³ania na zmiennych dwumianowych.
}
\details{
\tabular{ll}{
Package: \tab MDBinom\cr
Type: \tab Package\cr
Version: \tab 4.4\cr
Date: \tab 2012-08-01\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

Ró¿ne dzia³ania na zmiennych dwumianowych
}
\author{
Micha³ Danaj \email{michal.danaj@gmail.com}
}
\keyword{package}

