% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{model_dev}
\alias{model_dev}
\title{Odchylenie modelu}
\usage{
model_dev(predicted, target)
}
\arguments{
\item{predicted}{prawdopodobieństwo.}

\item{target}{target.}
}
\value{

}
\description{
Odchylenie modelu
}
\author{
Michał Danaj
}
