% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{na.subst}
\alias{na.subst}
\title{Zastêpuje braki danych}
\usage{
na.subst(x, val)
}
\arguments{
\item{x}{x}

\item{val}{jak¹ wartoœci¹ zast¹piæ brak danych.}
}
\value{

}
\description{
Zastêpuje braki danych
}
\author{
Piotr
}

