% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_i_sterujace.R
\docType{data}
\name{numeric_var_treatment.params}
\alias{numeric_var_treatment.params}
\title{Parametry}
\format{Wektor o polach:
\describe{
  \item{discret_threshold}{Przy jakiej wartoœci unikalnych wartoœci zmienna numeryczna bêdzie traktowana jako dyskretna.}
  \item{special_val}{Kody ró¿nych wartoœci specjalnych. Bêd¹ one traktowane jako wartoœci dyskretne, mimo ¿e pozosta³a czêœæ zmiennej mo¿e byæ ci¹g³a.}
}}
\usage{
numeric_var_treatment.params
}
\description{
Parametryzacja funkcji z pakietu.
}
\keyword{datasets}

