% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make_tree.r
\name{numeric_var_treatment}
\alias{numeric_var_treatment}
\title{Dyskretyzuje zmienną ciagłą drzewkiem w oparciu o zmienną odpowiedzi}
\usage{
numeric_var_treatment(x, y,
  special_val = numeric_var_treatment.params$special_val,
  NA_substit = numeric_var_treatment.params$NA_substit,
  min_bucket = numeric_var_treatment.params$min_bucket,
  max_gleb = numeric_var_treatment.params$max_gleb, interactive = FALSE,
  locfit = FALSE, breaks = NULL, span = 0.9, ...)
}
\arguments{
\item{x}{zmienna ciągła.}

\item{y}{zmienna celu (PD, LGD).}

\item{special_val}{wartości specjalne, które powinny zostać uwzględnione jako
osobne klasy. Wartość taką jest również \code{NA}, automatycznie
uwzględniana jako osobna klasa.}

\item{NA_substit}{wartość specjalna wstawiona za brak danych.}

\item{min_bucket}{Minimalna wielkość liścia}

\item{max_gleb}{Maksymalna głębokośc do której budujemy drzewo}

\item{interactive}{TRUE, jeśli zmienna ma być dyskretyzowana interaktywnie. W
przeciwnym razie, co jest wartością domyślną, dyskretyzacja
jest automatyczna.}

\item{locfit}{Czy z automatu dopasować funkcję z modelu \code{locfit}.}

\item{breaks}{Zamiast automatycznego dzielenia, można podać wartości przedziałów (from,to].}

\item{span}{Parametr wygładzający funkcji \code{locit}.}

\item{...}{inne parametry do funkcji \code{\link{drzewo}}.}
}
\description{
W parametrze \code{x} ani \code{y} nie może być NULLi. Można je jakoś zakodować.
}
\seealso{
\code{drzewo}
}
\author{
Michał Danaj
}
