% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{nvl}
\alias{nvl}
\title{Podstawia liczbę za brak danych}
\usage{
nvl(x, val)
}
\arguments{
\item{x}{objekt zmiennych numerycznych}

\item{val}{wartość, którą ma zostać zastąpiony brak danych \code{NA}}
}
\value{
Zwraca pierwotny obiekt z wartością \code{val} w miejsce braku danych
}
\description{
Podstawia liczbę za brak danych
}
\author{
Michał Danaj
}
