% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{nvl2}
\alias{nvl2}
\title{Zwraca jedną z dwóch wartości w zależności czy był brak danych}
\usage{
nvl2(x, val_not_null, val_null)
}
\arguments{
\item{x}{objekt zmiennych numerycznych}

\item{val_not_null}{wartość, która ma zostać zwrócona w przypadku gdy nie ma braku danych}

\item{val_null}{wartość, która ma zostać zwrócona w przypadku braku danych}
}
\value{
Zwraca pierwotny obiekt z oryginalnymi wartościami zastąpionymi przez \code{val_not_null} i \code{val_null}
}
\description{
Zwraca jedną z dwóch wartości w zależności czy był brak danych
}
\author{
Michał Danaj
}
