% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{plot_AR}
\alias{plot_AR}
\title{Rysuje krzywe ROC i CAP}
\usage{
plot_AR(ar, plot_type = c("ROC", "CAP"), adjusted_AR = FALSE, ...)
}
\arguments{
\item{ar}{Obiekt klasy XYZ.}

\item{plot_type}{Typ wykresu.}

\item{adjusted_AR}{zmienna logiczna. TRUE, gdy trzeba osobno wyliczyæ pole pod
modelem idealnym (np. w przypadku LGD).}
}
\description{
Dla obiektu zwróconego przez funkcjê \code{\link{AR}} rysuje krzyw¹ ROC lub CAP.
}
\details{
Dla obiektu zwróconego przez funkcjê \code{\link{AR}} rysuje krzyw¹ ROC lub CAP.
}
\author{
Micha³ Danaj
}

