% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{plotCalibr}
\alias{plotCalibr}
\title{Rysuje kalibracjê modelu}
\usage{
plotCalibr(score, response, estim, target = c("br", "logit"), ylab = y_name,
  xlab = "score", ...)
}
\arguments{
\item{score}{Zmienna na osi OX.}

\item{response}{Zmienna celu.}

\item{estim}{Wyestymowana wartoœæ przez model.}

\item{target}{Co ma byæ na osi OY.}

\item{ylab}{Opis osi OY.}

\item{xlab}{Opis osi OX.}

\item{...}{Parametry graficzne.}
}
\value{

}
\description{
Rysuje kalibracjê modelu
}
\author{
Piotr
}

