% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make.R
\name{polacz_buckety}
\alias{polacz_buckety}
\title{??czy ze sob? buckety}
\usage{
polacz_buckety(x, y, buckets, row_idxs, new_label = NULL)
}
\arguments{
\item{x}{zmienna score'owa.}

\item{y}{zmienna odpowiedzi z zakresu [0,1]. Np. default, LGD.}

\item{buckets}{buckety.}

\item{row_idxs}{numery wierszy w \code{buckets} do po??czenia.}

\item{new_label}{label jaki b?dzie nadany po??czonym bucketom. W przypadku braku podania, zostanie zatosowany domy?lny.}
}
\value{
\code{data.frame}.
}
\description{
??czy buckety. W przypadku bucket?w z przedzia?ami zmiennej ci?g?ej, mo?liwe jest 
po??czenie tylko
przedzia??w przylegaj?cych do siebie. Dla nowo powsta?ych bucket?w wylicza statystyki. 
W przypadku ??czenia przedzia??w zmiennej ci?g?ej, wiersze z tymi bucketami zostan? ze sob?
po??czone i powstanie \code{data.frame} z liczb? wierszy o jeden mniejsz? ni? w \code{bucket}.
Przy ??czeniu bucket?w dyskretnych lub dyskretnego i ci?g?ego, wiersze nie zostan? usuni?te. 
Zostanie im nadany wsp?lny label oraz wsp?lny numer. Je?li liczba bucket?w do po??czenia jest<=1,
zwracany jest wej?ciowy podzia?.
}
\author{
Micha? Danaj
}
