% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary.r
\name{prel_factors}
\alias{prel_factors}
\title{Podstawowe statystyki faktorów}
\usage{
prel_factors(dane)
}
\arguments{
\item{dane}{}
}
\value{

}
\description{
Wiele z faktorów zaczytanych bezpooerednio z pliku nie nadaje sie do modelowania. Funkcja umo?liwia wstepn^1 
analize faktorów.
}
\author{
Micha^3 Danaj
}
