% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary.R
\name{prel_numeric_one_value}
\alias{prel_numeric_one_value}
\title{Wyœwietla zmienne z wartoœci¹ o du¿ym udziale w rozk³adzie}
\usage{
prel_numeric_one_value(dane, rate = 0.98)
}
\arguments{
\item{dane}{Dane.}

\item{rate}{Graniczny udzia³ jedenej wartoœci.}
}
\value{

}
\description{
Wyœwietla nazwy zmiennych, dla których istenieje taka wartoœæ, ¿e jej udzia³ przekracza \code{rate}
próby.
}
\author{
Micha³ Danaj
}

