% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary.R
\name{prel_numeric_sd0}
\alias{prel_numeric_sd0}
\title{Wyœwietla zmienne numeryczne z wariancj¹ == 0}
\usage{
prel_numeric_sd0(dane)
}
\arguments{
\item{dane}{}
}
\value{

}
\description{
Wyœwietla zmienne numeryczne z jedn¹ wartoœci¹.
}
\author{
Micha³ Danaj
}

