% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_assign.R
\name{przypisz2}
\alias{przypisz2}
\title{Przypisuje dyskretyzacjê}
\usage{
przypisz2(x, bucket, interpol = FALSE, fitted = NULL,
  NA_substit = -2147483647)
}
\value{
Wektor wartoœci.
}
\description{
Przypisuje zmienne dyskretne oraz ci¹g³e w oparciu o podane przedzia³y, uwzglêdnia wartoœci
specjalne
  @param x Wektor zmiennych numerycznych. 
  @param bucket \code{data.frame} z bucketami.
  @param interpol Czy przeprowadziæ interpolacjê?
  @param NA_substit wartoœæ do podstawienia za brak danych
}
\author{
Micha³ Danaj
}
\seealso{
\code{\link{przypisz}}
}

