% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{przypisz_woe}
\alias{przypisz_woe}
\title{Przypisuje woe na podstawie br z bucketa.}
\usage{
przypisz_woe(bucket_list, data, vars = names(bucket_list),
  varname_sufix = "woe")
}
\arguments{
\item{bucket_list}{- lista z opisami dyskretyzacji. Nazwy elementów listy powinny byæ zgodne z nazwami zmiennych w \code{data}.}

\item{data}{- \code{data.frame} z oryginalnymi zmiennymi.}

\item{vars}{- wektor z nazwami zmiennych do ograniczenia.}

\item{varname_sufix}{- Sufix wynikowych zmiennych.}
}
\description{
Przypisuje woe na podstawie br z bucketa. Korzysta z funkcji \code{woe}, która 
w razie niewyst¹pienia w buckecie wartoœci z klasy 0 lub 1, przyjmuje wartoœæ 0.5
}

