% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{przypisz_woe_z_listy}
\alias{przypisz_woe_z_listy}
\title{Przypisuje woe na podstawie br z bucketa.}
\usage{
przypisz_woe_z_listy(
  bucket_list,
  data,
  vars = names(bucket_list),
  varname_sufix = "woe"
)
}
\arguments{
\item{bucket_list}{- lista z opisami dyskretyzacji. Nazwy elementów listy powinny być zgodne z nazwami zmiennych w \code{data}.}

\item{data}{- \code{data.frame} z oryginalnymi zmiennymi.}

\item{vars}{- wektor z nazwami zmiennych do ograniczenia.}

\item{varname_sufix}{- Sufix wynikowych zmiennych.}
}
\description{
Przypisuje woe na podstawie br z bucketa. Korzysta z funkcji \code{woe}, która 
w razie niewystąpienia w buckecie wartości z klasy 0 lub 1, przyjmuje wartość 0.5
}
