% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{przypisz_z_listy}
\alias{przypisz_z_listy}
\title{Na podstawie listy z bucketami przypisuje odpowiednie wartości}
\usage{
przypisz_z_listy(
  bucket_list,
  data,
  vars = names(bucket_list),
  colname = "fitted",
  varname_sufix = colname,
  sep_sufix = "_"
)
}
\arguments{
\item{bucket_list}{Lista bucketów.}

\item{data}{Dane.}

\item{vars}{Zmienne do przypisania.}

\item{colname}{Nazwa kolumny z bucketa, na podstawie którego należy dopisać wartości.}

\item{varname_sufix}{Sufiks do nazw nowych zmiennych.}

\item{sep_sufix}{Parametr umożliwiający zmianę separatora sufiksa. Domyślnie podkreślenie - została ustalona
w celu zachowania kompatybilności z wersją funkcji bez tego parametru.}
}
\value{

}
\description{
Na podsatwie listy \code{bucket_list} przypisuje zmiennych \code{vars} z tabeli \code{data}
wartości znajdujące się w kolumnie \code{fitted} bucketów. Wartości przypisuje do kolumn
o nazwie z połącznia nazw \code{vars} i \code{varname_sufix}.
}
\author{
Michał Danaj
}
