% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{to_excel}
\alias{to_excel}
\title{Wrzuca dane do schowka, do wklejenia do Excela}
\usage{
to_excel(table, row.names = FALSE)
}
\arguments{
\item{table}{\code{data.frame} do przekazania do Excela.}

\item{row.names}{czy przekazać również wiersze.}
}
\value{

}
\description{
Wrzuca dane do schowka, do wklejenia do Excela
}
\author{
Michał Danaj
}
