% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate_anal_stats1}
\alias{univariate_anal_stats1}
\title{Dyskretyzuje zmienną i wylicza na niej statystyki}
\usage{
univariate_anal_stats1(
  x,
  y,
  locfit = FALSE,
  discrete_threshold = numeric_var_treatment.params$discrete_threshold,
  NA_substit = numeric_var_treatment.params$NA_substit,
  special_val = numeric_var_treatment.params$special_val,
  no_stats_threshold = numeric_var_treatment.params$no_stats_threshold,
  max_gleb = 3,
  min_bucket = 200,
  interactive = FALSE,
  breaks = NULL,
  mapping = NULL,
  forceContinous = FALSE,
  span = 0.9,
  ...
)
}
\arguments{
\item{x}{zmienna, po której procedura będzie sortować.}

\item{y}{zmienna odpowiedzi.}

\item{locfit}{Czy z automatu dopasować funkcję z modelu \code{locfit}.}

\item{discrete_threshold}{jeśli liczba unikalnych wartości zmiennej jest nie większa
ta wartość, zmienna uznana jest za dyskretną i nie jest poddawana dyskretyzacji.}

\item{NA_substit}{wartość, którą zastąpić brak danych}

\item{special_val}{Wartości specjalne do usunięcia z automatycznego podziału. Będą traktowane jako zmienne
kategoryczne.}

\item{no_stats_threshold}{liczba unikalnych wartości zmiennej kategorycznej, powyżej której nie są generowane
statystyki. W przypadku przekroczenia, zwracany jest komunikat "Too many categorical levels".}

\item{max_gleb}{Maksymalna głębokośc do której budujemy drzewo}

\item{min_bucket}{Minimalna wielkość liścia}

\item{interactive}{TRUE, jeśli zmienna ma być dyskretyzowana interaktywnie. W
przeciwnym razie, co jest wartością domyślną, dyskretyzacja
jest automatyczna.}

\item{breaks}{zamiast automatycznego dzielenia, można podać wartości przedziałów (from,to].}

\item{mapping}{zamiast automatycznego dzielenia, można podać mapowanie.}

\item{forceContinous}{wymusza potraktowanie zmiennej jako ciągłą, mimo że liczba
unikalnych wartości jest mniejsza niż \code{discrete_threshold}.}

\item{span}{Parametr wygładzający funkcji \code{locit}.}

\item{...}{dodatkowe parametry graficzne.}
}
\description{
W przypadku, gdy liczba unikalnych wartości zmiennej jest <= \code{discrete_threshold}
lub zmienna nie jest zmienną numeryczną,
uznaje że zmienna jest dyskretna i jedynie wylicza dla niej statystyki. W przeciwnym
wypadku dyskretyzuje zmienną i wylicza statystyki.
}
\seealso{
\code{\link{buckety_stat}}.
}
