% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate_anal_stats1b}
\alias{univariate_anal_stats1b}
\title{Dyskretyzuje zmienn¹ i wylicza na niej statystyki}
\usage{
univariate_anal_stats1b(x, y, locfit = FALSE, discret_treshold = 15,
  special_val = numeric_var_treatment.params$spcial_val, max_gleb = 3,
  min_bucket = 200, interactive = FALSE, breaks = NULL, mapping = NULL,
  forceContinous = FALSE, span = 0.9, ...)
}
\arguments{
\item{discret_treshold}{jeœli liczba unikalnych wartoœci zmiennej jest nie wiêksza
ta wartoœæ, zmienna uznana jest za dyskretn¹ i nie jest poddawana dyskretyzacji.}

\item{interactive}{TRUE, jeœli zmienna ma byæ dyskretyzowana interaktywnie. W
przeciwnym razie, co jest wartoœci¹ domyœln¹, dyskretyzacja
jest automatyczna.}

\item{breaks}{zamiast automatycznego dzielenia, mo¿na podaæ wartoœci przedzia³ów (from,to].}

\item{forceContinous}{wymusza potraktowanie zmiennej jako ci¹g³¹, mimo ¿e liczba
unikalnych wartoœci jest mniejsza ni¿ \code{discret_treshold}.}
}
\description{
W przypadku, gdy liczba unikalnych wartoœci zmiennej jest <= \code{discret_treshold}
lub zmienna nie jest zmienn¹ numeryczn¹,
uznaje ¿e zmienna jest dyskretna i jedynie wylicza dla niej statystyki. W przeciwnym
wypadku dyskretyzuje zmienn¹ i wylicza statystyki.
}
\seealso{
\code{\link{buckety_stat}}.
}

