% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate_anal_stats2}
\alias{univariate_anal_stats2}
\title{Robi rozk³ad zmiennej po czasie (lub innym podziale)}
\usage{
univariate_anal_stats2(x_discr, y, czas, estim)
}
\arguments{
\item{x_discr}{zmienna objaœniaj¹ca.}

\item{y}{zmienna odpowiedzi.}

\item{czas}{Czas.}
}
\description{
Wylicza licznoœci dla ka¿dego poziomu \code{x_discr} oraz œredni¹ wartoœæ \code{y}
w podziale na zadane grupy czasowe \code{czas} (lub podzia³ innego typu).
}

