% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate_anal_stats3}
\alias{univariate_anal_stats3}
\title{Wylicza AR po czasie i w podziale na zadane próby}
\usage{
univariate_anal_stats3(score, y, czas, proby)
}
\arguments{
\item{score}{zmienna, po której procedura bêdzie sortowaæ.}

\item{y}{zmienna odpowiedzi.}

\item{czas}{podzia³ na okresy czasowe.}

\item{proby}{\code{data.frame}, w którym ka¿da kolumna jest wektorem logicznym,
zawieraj¹cym informacje, czy obserwacja nale¿y do danej próby.}
}
\description{
Wylicza AR po czasie i w podziale na zadane próby
}

