% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate_anal_stats4}
\alias{univariate_anal_stats4}
\title{Nie wiem co to robi}
\usage{
univariate_anal_stats4(
  dane,
  mapowanie,
  czas = lastDay(dane$reportingdate, unit = "quater"),
  ...
)
}
\arguments{
\item{dane}{dane}

\item{mapowanie}{mapowanie}

\item{czas}{czas}

\item{...}{dodatkowe parametry graficzne.}
}
\value{

}
\description{
Nie wiem co to robi
}
\author{
Michał Danaj
}
