% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_loop.R
\name{univariate_loop}
\alias{univariate_loop}
\title{Grupuje zmienne i generuje dla nich listê statystyk}
\usage{
univariate_loop(x_df, y, vsub_bool = rep(TRUE, ncol(x_df)),
  vsub_list = NULL, vsub_rola = NULL, proby = rep(TRUE, length(y)),
  czas = rep(1, length(y)), ...)
}
\arguments{
\item{x_df}{\code{data.frame} ze zmiennymi do anliz.}

\item{y}{\code{vector} z wartoœciami zmiennej odpowiedzi.}

\item{vsub_bool}{\code{vector} logiczny okreœlaj¹cy, które zmienne z \code{df} powinny zostaæ zanalizowane.}

\item{vsub_list}{\code{vector} z nazwami zmiennych z \code{df}, które powinny zostaæ zanalizowane.}

\item{vsub_rola}{\code{vector} z rol¹ ka¿dej ze zmiennych. Zostan¹ zanalizowane zmienne, dla których rola 
ustawiona jest na \code{'explanatory'}.}

\item{proby}{\code{data.frame} z wektorami logicznymi, okreœlaj¹cymi podpróby na których nale¿y przeanalizowaæ podpróby.}

\item{czas}{\code{vector} ze zmienn¹ czasow¹, po której zostan¹ podzielone analizy.}

\item{...}{dodatkowe parametry do funkcji \code{\link{univariate_anal_stats}}.}
}
\value{
lista ze statstykami dla ka¿dej ze zmiennych okreœlonych przez któryœ z wektorów vsub\*. Nazwy elementów listy
	s¹ nazwami zmiennych.
}
\description{
Funkcja grupuje zmienne ci¹g³e, zgodnie z domyœlnymi parametrami. Na grupach oraz na zmiennych kategorycznych 
wylicza statytystyki i przedstawia je w formie listy, osobno dla ka¿dej ze zmiennych. Wykorzystuje do tego funkcjê
\code{\link{univariate_anal_stats}}.
}
\author{
Michal Danaj
}
\seealso{
\code{\link{univariate_anal_stats}}
}

