% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate_stats_new_data}
\alias{univariate_stats_new_data}
\title{Wylicza statystyki dla zmiennej score'owej i objaśnianej wg zadanych bucketów}
\usage{
univariate_stats_new_data(buckets, x, y, czas, proby = rep(TRUE, length(y)))
}
\arguments{
\item{buckets}{\code{data.frame} z podziałem zmiennej na buckety.}

\item{x}{zmienna, po której procedura będzie sortować.}

\item{y}{zmienna odpowiedzi.}

\item{czas}{podział na okresy czasowe.}

\item{proby}{\code{data.frame}, w którym każda kolumna jest wektorem logicznym,
zawierającym informacje, czy obserwacja należy do danej próby.}
}
\description{
Wylicza statystyki i zwraca je w postaci listy. Do wyliczenia \code{AR} używa jako zmiennej
score'owej oryginalnych wartości BR przypisanych do bucketu. Tzn, mimo że na nowych danych kolejność
bucketów, sortując je po br może być inna niż oryginalnie, to stosowana jest oryginalna kolejność.
}
\author{
Michał Danaj
}

