% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{usun_konce}
\alias{usun_konce}
\title{Usuwa krańcowe wartości.}
\usage{
usun_konce(score, prob = 0.01, weights = NULL)
}
\arguments{
\item{score}{Wektor wartości numerycznych.}

\item{prob}{Jaką część obserwacj należy usunąć z jednego krańca.}

\item{weights}{Wagi.}
}
\value{
Zwraca wektor z indeksami elementów, które powinny zostać usunięte.
}
\description{
Zwraca indeksy \code{prob} najmniejszych i \code{prob} największych wartości. Jeśli
jednak nie jest możliwe ustalenie co najwyżej \code{prob} obserwacji, nie 
zostaje wybrana żadna. Ma to zapobiec sytuacji, gdy np. najmniejszą wartością jest 0
i w próbie znajduje się 10\% takich obserwacji, a my chcemy usunąć tylko 1\%.
}
\examples{
x<-sort(rnorm(10));
x
usun<-usun_konce(x, prob=0.01);
x[-usun]

#usuwa tylko obserwacjÄ™ z prawego kraĹ„ca
x2<-c(rep(min(x),5),x[5:10])
x2
usun<-usun_konce(x2, prob=0.01);
x2[-usun]
}
\author{
Michał Danaj
}
