% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{AR}
\alias{AR}
\title{Liczy Accuracy Ratio (GINI)}
\usage{
AR(score, def, plot = FALSE, sort.order = -1, label = "", ...)
}
\arguments{
\item{score}{wektor    wartości, według który należy posortować wartości \code{def}.}

\item{def}{Wektor wartości \code{\{0,1\}}.}

\item{plot}{Czy narysować krzywą CAP. Domyślnie \code{FALSE}.}

\item{sort.order}{Dla \code{sort.order=1} sortuje rosnąco, a dla \code{sort.order=-1} malejąco
po \code{score}.}

\item{label}{opis wyświetlony w legendzie.}

\item{...}{dodatkowe parametry.}
}
\value{
Zwraca wektor ze statystykami jakości modelu.
 \item{AUCAP}{Powierzchnia pod krzywą CAP.}
 \item{AR}{Wyliczone Accuracy Ratio (GINI).}
 \item{AUC}{Powierzchnia pod krzywą ROC.}							
 \item{KS}{Statystyka Kołmogorowa-Smirnowa.}
}
\description{
Wylicza Accuracy Ratio (GINI). Funkcja jest znacznie uproszczona w porównaniu z poprzednią wersją 
i funkcje nie są ze sobą kompatybilne. Największą zmianą jest wykorzystanie biblioteki \code{data.table},
dzięki czemu nieporównywalnie zwiększona została prędkość działania.
}
\author{
Michał Danaj
}
