% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{AR}
\alias{AR}
\title{Liczy Accuracy Ratio (GINI)}
\usage{
AR(score, def, plot = FALSE, return.table = FALSE, sort.order = NULL,
  label = "", lgd_adjusted = FALSE, ...)
}
\arguments{
\item{return.table}{czy zwracaæ tabelê z agregatami na poziomie pojedynczej 
        wartoœci score
 @param sort.order dla wartoœci równej "br", zamiast po \code{score}, 
             sortuje po wartoœci \code{br}. Mo¿e to byæ przydatne przy zmiennych dyskretnych.
 @param label opis wyœwietlony w legendzie.
 @param lgd_adjusted W przypadku modeli LGD pole pod krzywymi CAP i ROC dla modelu idealnego jest
            mniejsze ni¿ dla modelu PD i nie da siê go wyliczyæ analogicznym wzorem. Dlatego
dla LGD wyliczana jest osobno powierzchnia dla modelu idealnego i jest ona u¿yta jako mianownik do
wyliczenia AR. 
 @param ... dodatkowe parametry.}
}
\value{
Zwraca listê obiektów \code{data.frame}. Elementy tej listy odpowiadaj¹ kolejnym
kolumnom ze \code{score_wiele} i maj¹ takie same nazwy, jak kolumny ze \code{score_wiele}.
Ka¿dy element listy ma postaæ:
  \item{do }{Górna granica przedzia³u, dla którego podane s¹ statystyki.}
  \item{bad}{Liczba bad-ów w przedziale.}
  \item{obs}{Liczba obserwacji w przedziale.}
  \item{pct_all}{Procentowy udzia³ obserwacji w danym przedziale.}
  \item{pct_bad}{Procentowy udzia³ bad-ów w danym przedziale w stosunku do 
									wszystkich bad-ów.}
  \item{pct_good}{Procentowy udzia³ good-ów w danym przedziale w stosunku do 
									wszystkich good-ów.}
  \item{br}{Stosunek bad-ów do good-ów w danym przedziale (bad rate).}
  \item{AR}{Wyliczone Accuracy Ratio.}
}
\description{
Wylicza Accuracy Ratio (GINI). Wynikiem jest \code{data.frame} przechowuj¹cy
 informacje pozwalaj¹ce na wyrysowanie krzywej CAP. W ostatniej kolumnie zapiasna
 jest wartoœæ AR.
}
\details{
@param score wektor	wartoœci, wed³ug który nale¿y posortowaæ wartoœci \code{def}
 @param def Wektor wartoœci \code{\{0,1\}} 
 @param plot Czy narysowaæ krzyw¹ CAP. Domyœlnie \code{FALSE}.
}
\examples{
n<-1000;
x<-rnorm(n);
def<- as.numeric(x+rnorm(n)<0);
AR(x,def,plot=TRUE);
}
\author{
Micha³ Danaj
}

