% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{AR_quick}
\alias{AR_quick}
\title{Liczy AR(GINI) dla specyficznie przygotowanych danych}
\usage{
AR_quick(bad, obs)
}
\arguments{
\item{bad}{liczba badów odpowiadająca kolejnym wartościom score.}

\item{obs}{liczba obserwacji odpowiadająca kolejnym wartościom score.}
}
\description{
Funkcja zakłada, że podane liczności badów i wszystich obserwacji
posortowane są wg rosnących wartości score. W przypadku gdy w próbie są
tylko bady lub tylko goody, funkcja zwraca \code{NaN}.
}
\author{
Michał Danaj
}
