% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make.R
\name{bckt_stat2}
\alias{bckt_stat2}
\title{Wylicza statystyki dla podanych przedzia??w (dev)}
\usage{
bckt_stat2(breaks, x = NULL, y = NULL, weights = rep(1, length(x)),
  dt = NULL, avg = NULL, total = TRUE, sort_x = TRUE)
}
\arguments{
\item{breaks}{punkty podzia?u}

\item{x}{zmienna score'owa.}

\item{y}{zmienna odpowiedzi z zakresu [0,1]. Np. default, LGD.}

\item{weights}{wektor z wagami.}

\item{dt}{\code{data.table} z kolumnami \code{x}, \code{y}, \code{weights}.}

\item{avg}{\code{vector}, \code{data.frame}, \code{data.table} z warto?ciami dla kt?rych
zostan? wyliczone ?rednie.}

\item{total}{czy do??czy? wiersz z podsumowaniem.}

\item{sort_x}{je?li \code{TRUE}, wynikowa tabela zostanie posortowana po warto?ciach}
}
\value{
\code{data.frame} ze statystykami.
}
\description{
Dzieli zmienn? \code{x} na przedzia?y okre?lone przez \code{breaks}
i wylicza statystyki dla tak powsta?ych bucket?w.
}
\details{
Do przypisania warto?ci do przedzia?u u?ywana jest funkcja

\code{findInterval(x, y, rightmost.closed = TRUE, all.inside = TRUE)}.
}
\seealso{
\code{\link{buckety_stat}}.
}
\author{
Micha? Danaj
}
