% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make.r
\name{buckety_br}
\alias{buckety_br}
\title{Dzieli na przedziały jedną z dwóch metod i wylicza na nich bad rate.}
\usage{
buckety_br(x, y, n, method = c("eq_length", "eq_count"),
  one.value.action = c("none", "combine"), weights = NULL)
}
\arguments{
\item{x}{Zmienna, którą będziemy dyskretyzować.}

\item{y}{Zmienna dwumianowa.}

\item{n}{Liczba wynikowych przedziałów.}

\item{method}{Sposób podziału na przedziały. Szczegóły w sekcji Details.}

\item{one.value.action}{Jeszcze nie działa.}

\item{weights}{Wagi.}
}
\value{
Zwraca \code{data.frame}, w którym dla \code{i}-tego wiersza podane są statystyki
dla \code{i}-tego przedziału.
\item{nr }{Numer przedziału.}
\item{od }{Początek przedziału.}
\item{srodek}{Środek przedziału.}
\item{mean}{Średnia z obserwacji w przedziale.}
\item{do}{Koniec przedziału.}
\item{n_default}{Liczba defaultów.}
\item{n_obs}{Liczba obserwacji.}
\item{br}{Bad rate.}
\item{logit}{Logit.}
\item{probit}{Probit.}
\item{var}{Wariancja zmiennej \code{default}.}
\item{waga}{Waga jako odwrotność warinacji. Gdy \code{n_default=0} to przyjmuje się,
	że \code{n_default=0.5}}
}
\description{
Jako że jest to wersja przejściowa, nie ma żadnych detali.
Może poza tym, że jeśli nie ma obserwacji w jakimś bukcecie, to go usuwa.
}
\examples{
x<-rnorm(1000);
y<-(x+rnorm(1000))<0;
buckety_br(x, y, 10);
}
\author{
Michał Danaj
}
