% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make.R
\name{buckety_br}
\alias{buckety_br}
\title{Dzieli na przedzia?y jedn? z dw?ch metod i wylicza na nich bad rate.}
\usage{
buckety_br(x, y, n, method = c("eq_length", "eq_count"),
  one.value.action = c("none", "combine"), weights = NULL)
}
\arguments{
\item{x}{Zmienna, kt?r? b?dziemy dyskretyzowa?.}

\item{y}{Zmienna dwumianowa.}

\item{n}{Liczba wynikowych przedzia??w.}

\item{method}{Spos?b podzia?u na przedzia?y. Szczeg??y w sekcji Details.}

\item{one.value.action}{Jeszcze nie dzia?a.}

\item{weights}{Wagi.}
}
\value{
Zwraca \code{data.frame}, w kt?rym dla \code{i}-tego wiersza podane s? statystyki
dla \code{i}-tego przedzia?u.
\item{nr }{Numer przedzia?u.}
\item{od }{Pocz?tek przedzia?u.}
\item{srodek}{?rodek przedzia?u.}
\item{mean}{?rednia z obserwacji w przedziale.}
\item{do}{Koniec przedzia?u.}
\item{n_default}{Liczba default?w.}
\item{n_obs}{Liczba obserwacji.}
\item{br}{Bad rate.}
\item{logit}{Logit.}
\item{probit}{Probit.}
\item{var}{Wariancja zmiennej \code{default}.}
\item{waga}{Waga jako odwrotno?? warinacji. Gdy \code{n_default=0} to przyjmuje si?,
	?e \code{n_default=0.5}}
}
\description{
Jako ?e jest to wersja przej?ciowa, nie ma ?adnych detali.
Mo?e poza tym, ?e je?li nie ma obserwacji w jakim? bukcecie, to go usuwa.
}
\examples{
x<-rnorm(1000);
y<-(x+rnorm(1000))<0;
buckety_br(x, y, 10);
}
\author{
Micha? Danaj
}
