% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make.r
\name{buckety_stat}
\alias{buckety_stat}
\title{Wylicza statystyki zmiennej \code{default} dla podanych grup \code{buckets}}
\usage{
buckety_stat(bucket, default, total=TRUE)
}
\arguments{
\item{bucket}{wektor z identyfikatorami bucketów.}

\item{default}{wektor zmiennych numerycznych.}

\item{total}{czy dołączyć wiersz z podsumowaniem.}
}
\value{
Zwraca \code{data.frame}, w którym dla \code{i}-tego wiersza podane są statystyki
  dla \code{i}-tego przedziału.
  \item{nr }{Numer przedziału.}
  \item{od }{Początek przedziału.}
  \item{srodek}{Środek przedziału.}
  \item{mean}{Średnia z obserwacji w przedziale.}
  \item{do}{Koniec przedziału.}
  \item{n_default}{Liczba defaultów.}
  \item{n_obs}{Liczba obserwacji.}
  \item{br}{Bad rate.}
  \item{logit}{Logit.}
  \item{probit}{Probit.}
  \item{var}{Wariancja zmiennej \code{default}.}
  \item{waga}{Waga jako odwrotność warinacji. Gdy \code{n_default=0} to przyjmuje się,
			że \code{n_default=0.5}}
}
\description{
Wylicza statystyki zmiennej \code{default} dla podanych grup \code{buckets}
}
\author{
Michał Danaj
}
