% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make.R
\name{buckety_stat}
\alias{buckety_stat}
\title{Wylicza statystyki zmiennej \code{default} dla podanych grup \code{buckets}}
\usage{
buckety_stat(bucket, default, total=TRUE)
}
\arguments{
\item{bucket}{wektor z identyfikatorami bucketów.}

\item{default}{wektor zmiennych numerycznych.}

\item{total}{czy do³¹czyæ wiersz z podsumowaniem.}
}
\value{
Zwraca \code{data.frame}, w którym dla \code{i}-tego wiersza podane s¹ statystyki
  dla \code{i}-tego przedzia³u.
  \item{nr }{Numer przedzia³u.}
  \item{od }{Pocz¹tek przedzia³u.}
  \item{srodek}{Œrodek przedzia³u.}
  \item{mean}{Œrednia z obserwacji w przedziale.}
  \item{do}{Koniec przedzia³u.}
  \item{n_default}{Liczba defaultów.}
  \item{n_obs}{Liczba obserwacji.}
  \item{br}{Bad rate.}
  \item{logit}{Logit.}
  \item{probit}{Probit.}
  \item{var}{Wariancja zmiennej \code{default}.}
  \item{waga}{Waga jako odwrotnoœæ warinacji. Gdy \code{n_default=0} to przyjmuje siê,
			¿e \code{n_default=0.5}}
}
\description{
Wylicza statystyki zmiennej \code{default} dla podanych grup \code{buckets}
}
\author{
Micha³ Danaj
}

