% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make.R
\name{buckety_stat2}
\alias{buckety_stat2}
\title{Wylicza statystyki dla podanych przedzia??w}
\usage{
buckety_stat2(breaks, score, def, total = FALSE)
}
\arguments{
\item{breaks}{punkty podzia?u}

\item{score}{zmienna score'owa.}

\item{def}{zmienna odpowiedzi z zakresu [0,1]. Np. default, LGD.}

\item{total}{czy wylicza? wiersz z podsumowaniem.}
}
\value{
\code{data.frame} ze statystykami.
}
\description{
Dzieli zmienn? \code{score} na przedzia?y okre?lone przez \code{breaks}
i wylicza statystyki dla tak powsta?ych bucket?w przy pomocy funkcji.
}
\details{
Do przypisania warto?ci do przedzia?u u?ywana jest funkcja

\code{findInterval(score, breaks, rightmost.closed = TRUE, all.inside = TRUE)}.
}
\seealso{
\code{\link{buckety_stat}}.
}
\author{
Micha? Danaj
}
