% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make.R
\name{buckety_stat2}
\alias{buckety_stat2}
\title{Wylicza statystyki dla podanych przedzia³ów}
\usage{
buckety_stat2(breaks, score, def, total = FALSE)
}
\arguments{
\item{breaks}{punkty podzia³u}

\item{score}{zmienna score'owa.}

\item{def}{zmienna odpowiedzi z zakresu [0,1]. Np. default, LGD.}

\item{total}{czy wyliczaæ wiersz z podsumowaniem.}
}
\value{
\code{data.frame} ze statystykami.
}
\description{
Dzieli zmienn¹ \code{score} na przedzia³y okreœlone przez \code{breaks}
i wylicza statystyki dla tak powsta³ych bucketów przy pomocy funkcji.
}
\details{
Do przypisania wartoœci do przedzia³u u¿ywana jest funkcja

\code{findInterval(score, breaks, rightmost.closed = TRUE, all.inside = TRUE)}.
}
\author{
Micha³ Danaj
}
\seealso{
\code{\link{buckety_stat}}.
}

