% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make.r
\name{buckety_stat2}
\alias{buckety_stat2}
\title{Wylicza statystyki dla podanych przedziałów}
\usage{
buckety_stat2(breaks, score, def, total = FALSE)
}
\arguments{
\item{breaks}{punkty podziału}

\item{score}{zmienna score'owa.}

\item{def}{zmienna odpowiedzi z zakresu [0,1]. Np. default, LGD.}

\item{total}{czy wyliczać wiersz z podsumowaniem.}
}
\value{
\code{data.frame} ze statystykami.
}
\description{
Dzieli zmienną \code{score} na przedziały określone przez \code{breaks}
i wylicza statystyki dla tak powstałych bucketów przy pomocy funkcji.
}
\details{
Do przypisania wartości do przedziału używana jest funkcja

\code{findInterval(score, breaks, rightmost.closed = TRUE, all.inside = TRUE)}.
}
\seealso{
\code{\link{buckety_stat}}.
}
\author{
Michał Danaj
}
