% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make_tree.R
\name{drzewo}
\alias{drzewo}
\title{Automatyczna dyskretyzacja w oparciu o algorytm drzewa}
\usage{
drzewo(score, def, freq = NULL, wytnij = 0, min_split = 30,
  min_bucket = 10, max_gleb = 4, n_buckets = 20, plot = TRUE,
  testy = FALSE, ...)
}
\arguments{
\item{score}{Zmienna do dyskretyzacji}

\item{def}{Zmienna celu}

\item{freq}{Licznoœæ. Na razie nie obs³ugiwane}

\item{wytnij}{Jak¹ czêœæ koñcowych wartoœci usun¹æ funkcj¹ \code{usun_konce}}

\item{min_split}{Minimalna wielkoœæ wêz³a, aby móc dokonaæ podzia³u}

\item{min_bucket}{Minimalna wielkoœæ liœcia}

\item{max_gleb}{Maksymalna g³êbokoœc do której budujemy drzewo}

\item{n_buckets}{Chyba nie u¿ywane?}

\item{plot}{Rysuje wynik dyskretyzacji}

\item{testy}{Czy wyœwietlaæ komentarze do testowania}

\item{...}{Dodatkowe parametry do rysowania}
}
\value{

}
\description{
Dyskretyzuje zmienn¹ \code{score}
}
\author{
Micha³ Danaj
}

