% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make_tree.R
\name{drzewo_plot}
\alias{drzewo_plot}
\title{Rysuje dyskretyzacjê w oparciu o drzewko}
\usage{
drzewo_plot(liscie_drzewa, ...)
}
\arguments{
\item{liscie_drzewa}{Cholera wie, co jest i jak mam to tu przekazaæ.}

\item{...}{Dodatkowe parametry graficzne.}
}
\value{

}
\description{
Rysuje dyskretyzacjê w oparciu o drzewko
}
\author{
Piotr
}

