% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make_tree.R
\name{drzewo_podzial}
\alias{drzewo_podzial}
\title{Podzia³ drzewa}
\usage{
drzewo_podzial(score, def, nr_wezla, od, do, freq, glebokosc, min_split = 200,
  min_bucket = 100, max_gleb = 3, testy = FALSE)
}
\arguments{
\item{score}{score}

\item{def}{def}

\item{nr_wezla}{nr_wezla}

\item{od}{od}

\item{do}{do}

\item{freq}{freq}

\item{glebokosc}{glebokosc}

\item{min_split}{min_split}

\item{min_bucket}{min_bucket}

\item{max_gleb}{max_gleb}

\item{testy}{testy}
}
\value{

}
\description{
Podzia³ drzewa.
}
\author{
Piotr
}

