% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_SQL.R
\name{genSQLCases}
\alias{genSQLCases}
\title{Generujê listê kodów SQL z dyskretyzacj¹ wszystkich zmiennych z \code{wynik}..}
\usage{
genSQLCases(wynik, dane, varNames)
}
\arguments{
\item{dane}{dane ze zmiennymi z modelu do okreœlenia typu danych.}

\item{varNames}{\code{vector} z nazwami zmiennych wynikowych.}

\item{wyniki}{wyniki dyskretyzacji.}
}
\value{
zwraca listê z kodami dla poszczególnych zmiennych.
}
\description{
Generujê listê kodów SQL z dyskretyzacj¹ wszystkich zmiennych z \code{wynik}.
}
\author{
Piotr
}

