% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_SQL.R
\name{genSQLCases2}
\alias{genSQLCases2}
\title{Generujê listê kodów SQL z dyskretyzacj¹ wszystkich zmiennych z \code{wynik}..}
\usage{
genSQLCases2(wynik, dane, mapping)
}
\arguments{
\item{dane}{dane ze zmiennymi z modelu do okreœlenia typu danych.}

\item{mapping}{mapowanie nazwy Ÿród³owej (przed zastosowniem dyskretyzacji z \code{wynik}
    na nazwê, która jest póŸniej w modelu. Jeœli jedna i druga nazwa jest taka sama,
to znaczy ¿e nie by³o robionej dyskretyzacji i case jest dla tej zmeinnej
pominiêty.}

\item{wyniki}{wyniki dyskretyzacji.}
}
\value{
zwraca listê z kodami dla poszczególnych zmiennych.
}
\description{
Generujê listê kodów SQL z dyskretyzacj¹ wszystkich zmiennych z \code{wynik}.
}
\author{
Piotr
}

