% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_SQL.R
\name{genSQLModel}
\alias{genSQLModel}
\title{Kod SQL modelu logistycznego.}
\usage{
genSQLModel(model, varNames = NULL)
}
\arguments{
\item{model}{model logistyczny.}

\item{varNames}{alternatywne nazwy zmiennych. Jeœli nie podane, to pobrane z \code{model}.}
}
\value{
Kod SQL.
}
\description{
Kod SQL modelu logistycznego.
}
\author{
Micha³ Danaj
}

