% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{informacje}
\alias{informacje}
\title{Rysuje informacje o danych}
\usage{
informacje(score, default, buckets = 20, span = 0.8, main = "",
  hist_col = "blue", ylab = c("Frequency", "BR"),
  xlab = "Calibrated score", ...)
}
\arguments{
\item{score}{Wektor ze zmienn¹ numeryczn¹.}

\item{default}{Wektor ze zmienn¹ dwumianow¹.}

\item{buckets}{Sugerowana liczba bucketów.}

\item{span}{Wspó³czynnik wyg³adzaj¹cy, wykorzystywany przez funkcjê \code{\link[locfit]{locfit}}}

\item{main}{Tytu³ wykresu.}

\item{hist_col}{Kolor histogramu}

\item{ylab}{Label}

\item{xlab}{Label}

\item{...}{dodatkowe parametry.}
}
\description{
Rysuje histogram zmiennej \code{score}, dla ka¿dego przedzia³u histogramu
 rysuje bad rate, oraz rysujê krzyw¹ regresji dla \code{default~score}. Pod
 wykresem wyœwietla informacje o Accuracy Ratios i bad rate.
}
\examples{
x<-rnorm(100);
y<-(x+rnorm(100))>0;
informacje(x, y, buckets=15, span=0.8, main="", hist_col="blue", type="br", method_bucket="eq_length")
}
\author{
Micha³ Danaj
}

