% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{informacje_kal}
\alias{informacje_kal}
\title{Rysuje jakoœæ kalibracji}
\usage{
informacje_kal(score, default, estym, buckets = 20, span = 0.8,
  hist_col = "blue", ylab = c("Frequency", "BR"),
  xlab = "Calibrated score", legx = 0, legy = 0, legcex = 1,
  leg.label = quote(c("Observations frequency", "Average LGD",
  "Smoothed average LGD", "Expected LGD")), ylim2 = NULL, ...)
}
\arguments{
\item{score}{score.}

\item{default}{default.}

\item{estym}{wartoœci wyestymowane przez model.}

\item{buckets}{Sugerowana liczba bucketów.}

\item{span}{Wspó³czynnik wyg³adzaj¹cy, wykorzystywany przez funkcjê \code{\link[locfit]{locfit}}}

\item{hist_col}{Kolor histogramu}

\item{ylab}{Label}

\item{xlab}{Label}

\item{legx}{po³o¿enie legendy}

\item{legy}{po³o¿enie legendy}

\item{legcex}{po³o¿enie legendy.}

\item{leg.label}{etykiety do legendy.}

\item{ylim2}{ograniczenie y-ka.}

\item{...}{dodatkowe parametry do funkcji \code{\link{plot}}.}
}
\description{
Rysuje jakoœæ kalibracji
}

