% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interactionsTest}
\alias{interactionsTest}
\title{Testuje interakcje między dwoma zmiennymi}
\usage{
interactionsTest(var1, var2, target)
}
\arguments{
\item{var1}{wektor z wartościami pierwszej zmiennej.}

\item{var2}{wektor z wartościami drugiej zmiennej.}

\item{target}{wektor z wartościami zmiennej celu.}
}
\value{
Zwraca listę z szeregiem statystyk:
\item{"p-value"}{p-value interakcji} 
\item{licznosc}{liczności obserwacji na przecięciach zmiennych}
\item{licznosc_target}{liczności targetu na przecięciach zmiennych}
\item{mean_real}{średnia wartość targetu  na przecięciach zmiennych}
\item{mean_pred1}{predykcja z modelu bez interakcji na przecięciach zmiennych} 
\item{mean_pred2}{precykcja z modelu z interakcjami na przecięciach zmiennych}
}
\description{
Funkcja zakłada, że \code{var1} oraz \code{var2} są zmiennymi dyskretnymi. Konwertowane są one do
typu \code{factor} i sprawdzany jest p-value modelu z interakcjami w stosunku do modelu bez interakcji.
Wykorzystany jest test Chi-kwadrat odchylenia modelu.
}
\author{
Michał Danaj
}
