% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interactionsTestAll}
\alias{interactionsTestAll}
\title{Testuje interakcje pomiędzy wszystkimi zmiennymi}
\usage{
interactionsTestAll(
  data,
  discr_list,
  target = data$target,
  vars_names = names(discr_list),
  colname = "label"
)
}
\arguments{
\item{data}{\code{data.frame} z danymi.}

\item{discr_list}{lista z opisem dyskretyzacji.}

\item{target}{wektor z wartościami zmiennej celu}

\item{vars_names}{wektor nazw zmiennych do sprawdzenia interakcji. Domyślnie wszystkie zmienne z \code{discr_list}.}

\item{colname}{nazwa kolumny z tabeli z definicją dyskretyzacji z której zostaną pobrane wartości do
przypisania dyskretyzacji zmiennej.}
}
\value{
Zwrócona zostanie lista z dwoma elementami:
\item{pvalue_table}{Zawiera \code{data.frame} z nazwą pierwszej zmiennej, drugiej zmiennej, oraz p-value interakcji.}
\item{details}{Zawiera listę ze statystykami zwróconymi przez \code{\link{interactionsTest}}.}
}
\description{
Na podstawie zmiennych źródłowych oraz definicji
dyskretyzacji z \code{discr_list}, funkcja przypisuje zmiennym ich wartości dyskretne oraz
testuje interakcję każdej pary ze zmiennych określonych w \code{vars_names}. Do testowania korzysta z funkcji \code{\link{interactionsTest}}.
}
\author{
Michał Danaj
}
