% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make_tree.R
\name{interactive_tree}
\alias{interactive_tree}
\title{Interaktywny podzia³ zmiennej ci¹g³ej na buckety}
\usage{
interactive_tree(score, def, span = 0.8, min_split = 200,
  min_bucket = 100, buckets = 60, max_gleb = 2)
}
\arguments{
\item{score}{zmienna score'owa.}

\item{def}{zmienna odpowiedzi z zakresu [0,1]. Np. default, LGD.}

\item{span}{Parametr wyg³adzaj¹cy funkcji \code{locit}.}

\item{min_split}{Minimalna wielkoœæ wêz³a, aby móc dokonaæ podzia³u}

\item{min_bucket}{Minimalna wielkoœæ liœcia}

\item{buckets}{Chyba nie u¿ywane?}

\item{max_gleb}{Maksymalna g³êbokoœc do której budujemy drzewo}
}
\value{
\code{data.frame} ze statystykami.
}
\description{
Interaktywny podzia³ zmiennej ci¹g³ej na buckety. Uwaga! W danych
nie mo¿e byæ wartoœci NULL. Nieopisane zmienne s¹ to zmienne z
wykorzystywanej funkcji \code{\link{drzewo}} oraz \code{\link{reg_nieparam}}.
Po wykonaniu zmiany wyœwietlane s¹ statystyki po bucketach oraz statystyka AR.
}
\author{
Micha³ Danaj
}
\seealso{
\code{\link{drzewo}}, \code{\link{AR_quick}},  \code{\link{buckety_stat2}}.
}

