% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.r
\name{korelacje_zmiennych}
\alias{korelacje_zmiennych}
\title{Wyświetla w gridzie wartości korelacji zmiennych z modelu}
\usage{
korelacje_zmiennych(model, data)
}
\arguments{
\item{model}{dopasowany model.}

\item{data}{dane.}
}
\value{

}
\description{
Wyświetla w gridzie wartości korelacji zmiennych z modelu
}
\author{
Michał Dnaaj
}
