% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{lastDay}
\alias{lastDay}
\title{Zwraca ostatni dzieñ podanego okresu.}
\usage{
lastDay(dat, unit = c("month", "quarter", "half_year", "year"), ...)
}
\arguments{
\item{dat}{Obiekt klasy \link{Date}, lub daj¹cy siê skonwertowaæ funkcj¹ \link{as.Date}.}

\item{unit}{Jednostka czasu, która ma byæ uwzglêdniona przy wyznaczaniu ostatniego dnia.
Mo¿liwe wartoœci: \code{c("month","quater", "year")}, które odpowiadaj¹ ostatniemu dniu miesi¹ca, kwarta³u, roku,
do którego nale¿y data \code{dat}.}

\item{...}{Opcje do funkcji \link{as.Date}.}
}
\description{
Zwraca ostatni dzieñ podanego okresu.
}
\author{
Micha³ Danaj
}

