% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make_model_formula}
\alias{make_model_formula}
\title{Na podstawie nazwy zmiennej celu i zmiennych objaœniaj¹cych, tworzy formu³ê modelu}
\usage{
make_model_formula(target, vars)
}
\arguments{
\item{target}{zmienna celu}

\item{vars}{zmienne objaœniaj¹ce}
}
\value{
forum³a do zastosowania w modelu
}
\description{
Na podstawie nazwy zmiennej celu i zmiennych objaœniaj¹cych, tworzy formu³ê modelu
}
\author{
Micha³ Danaj
}

