% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.r
\name{makeCoarseClassingTables}
\alias{makeCoarseClassingTables}
\title{Generuje tabelę z coarse classing do dokumentacji}
\usage{
makeCoarseClassingTables(wyniki)
}
\arguments{
\item{wyniki}{lista z wynikami dyskretyzacji}
}
\description{
Jeśli Woe lub IV wychochą +-Inf, to wstawia zamiast tego NA
}
