% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{na.subst}
\alias{na.subst}
\title{Zastępuje braki danych}
\usage{
na.subst(x, val)
}
\arguments{
\item{x}{x}

\item{val}{jaką wartością zastąpić brak danych.}
}
\value{

}
\description{
Zastępuje braki danych
}
\author{
Michał Danaj
}
