% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_i_sterujace.r
\docType{data}
\name{numeric_var_treatment.params}
\alias{numeric_var_treatment.params}
\title{Parametry}
\format{Wektor o polach:
\describe{
  \item{discret_threshold}{Przy jakiej wartości unikalnych wartości zmienna numeryczna będzie traktowana jako dyskretna.}
  \item{special_val}{Kody różnych wartości specjalnych. Będą one traktowane jako wartości dyskretne, mimo że pozostała część zmiennej może być ciągła.}
}}
\usage{
numeric_var_treatment.params
}
\description{
Parametryzacja funkcji z pakietu.
}
\keyword{datasets}
