% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make_tree.R
\name{numeric_var_treatment}
\alias{numeric_var_treatment}
\title{Dyskretyzuje zmienn¹ ciag³¹ drzewkiem w oparciu o zmienn¹ odpowiedzi}
\usage{
numeric_var_treatment(x, y,
  special_val = numeric_var_treatment.params$spcial_val,
  min_bucket = numeric_var_treatment.params$min_bucket,
  max_gleb = numeric_var_treatment.params$max_gleb, interactive = FALSE,
  locfit = FALSE, breaks = NULL, span = 0.9, ...)
}
\arguments{
\item{x}{zmienna ci¹g³a.}

\item{y}{zmienna celu (PD, LGD).}

\item{special_val}{wartoœci specjalne, które powinny zostaæ uwzglêdnione jako
osobne klasy. Wartoœæ tak¹ jest równie¿ \code{NA}, automatycznie
uwzglêdniana jako osobna klasa.}

\item{min_bucket}{Minimalna wielkoœæ liœcia}

\item{max_gleb}{Maksymalna g³êbokoœc do której budujemy drzewo}

\item{interactive}{TRUE, jeœli zmienna ma byæ dyskretyzowana interaktywnie. W
przeciwnym razie, co jest wartoœci¹ domyœln¹, dyskretyzacja
jest automatyczna.}

\item{locfit}{Czy z automatu dopasowaæ funkcjê z modelu \code{locfit}.}

\item{breaks}{Zamiast automatycznego dzielenia, mo¿na podaæ wartoœci przedzia³ów (from,to].}

\item{...}{inne parametry do funkcji \code{\link{drzewo}}.}

\item{Parametr}{wyg³adzaj¹cy funkcji \code{locit}.}
}
\description{
W parametrze \code{x} ani \code{y} nie mo¿e byæ NULLi. Mo¿na je jakoœ zakodowaæ.
}
\author{
Micha³ Danaj
}
\seealso{
\code{drzewo}
}

