% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{nvl}
\alias{nvl}
\title{Podstawia liczbê za brak danych}
\usage{
nvl(x, val)
}
\arguments{
\item{x}{objekt zmiennych numerycznych}

\item{val}{wartoœæ, któr¹ ma zostaæ zast¹piony brak danych \code{NA}}
}
\value{
Zwraca pierwotny obiekt z wartoœci¹ \code{val} w miejsce braku danych
}
\description{
Podstawia liczbê za brak danych
}
\author{
Micha³ Danaj
}

