% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{plotCalibr}
\alias{plotCalibr}
\title{Rysuje kalibracjê modelu}
\usage{
plotCalibr(score, response, estim, plt_type = c("br", "logit"),
  ylab = plt_type, xlab = "score", ...)
}
\arguments{
\item{score}{Zmienna na osi OX.}

\item{response}{Zmienna celu.}

\item{estim}{Wyestymowane wartoœæ przez model(e). Mo¿e byœ wektor lub lista wektorów.}

\item{plt_type}{Co ma byæ na osi OY. Czy œrednia z wartoœci \code{response}, czy logit.}

\item{ylab}{Opis osi OY.}

\item{xlab}{Opis osi OX.}

\item{...}{Parametry graficzne.}
}
\value{

}
\description{
Rysuje kalibracjê modelu
}
\author{
Micha³ Danaj
}

