% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.r
\name{plotCalibr}
\alias{plotCalibr}
\title{Rysuje kalibrację modelu}
\usage{
plotCalibr(score, response, estim, plt_type = c("br", "logit"),
  ylab = plt_type, xlab = "score", ...)
}
\arguments{
\item{score}{Zmienna na osi OX.}

\item{response}{Zmienna celu.}

\item{estim}{Wyestymowane wartość przez model(e). Może byś wektor lub lista wektorów.}

\item{plt_type}{Co ma być na osi OY. Czy średnia z wartości \code{response}, czy logit.}

\item{ylab}{Opis osi OY.}

\item{xlab}{Opis osi OX.}

\item{...}{Parametry graficzne.}
}
\value{

}
\description{
Rysuje kalibrację modelu
}
\author{
Michał Danaj
}
