% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_make.r
\name{polacz_buckety}
\alias{polacz_buckety}
\title{Łączy ze sobą buckety}
\usage{
polacz_buckety(x, y, buckets, row_idxs, new_label = NULL)
}
\arguments{
\item{x}{zmienna score'owa.}

\item{y}{zmienna odpowiedzi z zakresu [0,1]. Np. default, LGD.}

\item{buckets}{buckety.}

\item{row_idxs}{numery wierszy w \code{buckets} do połączenia.}

\item{new_label}{label jaki będzie nadany połączonym bucketom. W przypadku braku podania, zostanie zatosowany domyślny.}
}
\value{
\code{data.frame}.
}
\description{
Łączy buckety. W przypadku bucketów z przedziałami zmiennej ciągłej, możliwe jest 
połączenie tylko
przedziałów przylegających do siebie. Dla nowo powstałych bucketów wylicza statystyki. 
W przypadku łączenia przedziałów zmiennej ciągłej, wiersze z tymi bucketami zostaną ze sobą
połączone i powstanie \code{data.frame} z liczbą wierszy o jeden mniejszą niż w \code{bucket}.
Przy łączeniu bucketów dyskretnych lub dyskretnego i ciągłego, wiersze nie zostaną usunięte. 
Zostanie im nadany wspólny label oraz wspólny numer. Jeśli liczba bucketów do połączenia jest<=1,
zwracany jest wejściowy podział.
}
\author{
Michał Danaj
}
