% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdclassing.R
\name{polacz_buckety}
\alias{polacz_buckety}
\title{£¹czy ze sob¹ buckety}
\usage{
polacz_buckety(x, y, buckets, row_idxs, new_label = NULL)
}
\arguments{
\item{x}{zmienna score'owa.}

\item{y}{zmienna odpowiedzi z zakresu [0,1]. Np. default, LGD.}

\item{buckets}{buckety.}

\item{new_label}{label jaki bêdzie nadany po³¹czonym bucketom. W przypadku braku podania, zostanie zatosowany domyœlny.}
}
\author{
Micha³ Danaj
}

