% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_assign.R
\name{polacz_buckety}
\alias{polacz_buckety}
\title{£¹czy ze sob¹ buckety}
\usage{
polacz_buckety(x, y, buckets, row_idxs, new_label = NULL)
}
\arguments{
\item{x}{zmienna score'owa.}

\item{y}{zmienna odpowiedzi z zakresu [0,1]. Np. default, LGD.}

\item{buckets}{buckety.}

\item{new_label}{label jaki bêdzie nadany po³¹czonym bucketom. W przypadku braku podania, zostanie zatosowany domyœlny.}

\item{nr1}{numer wiersza w \code{buckets} z pierwszym bucketem do po³¹czenia.}

\item{nr2}{numer wiersza w \code{buckets} z pierwszym bucketem do po³¹czenia.}
}
\value{
\code{data.frame}.
}
\description{
£¹czy buckety. W przypadku bucketów z przedzia³ami zmiennej ci¹g³ej, mo¿liwe jest 
po³¹czenie tylko
przedzia³ów przlegaj¹cych do siebie. Dla nowo powsta³ych bucketów wylicza statystyki. 
W przypadku ³¹czenia przedzia³ów zmiennej ci¹g³ej, wiersze z tymi bucketami zostan¹ ze sob¹
po³¹czone i powstanie \code{data.frame} z liczb¹ wierszy o jeden mniejsz¹ ni¿ w \code{bucket}.
Przy ³¹czeniu bucketów dyskretnych lub dyskretnego i ci¹g³ego, wiersze nie zostan¹ usuniête. 
Zostanie im nadany wspólny label oraz wspólny numer. Jeœli liczba bucketów do po³¹czenia jest<=1,
zwracany jest wejœciowy podzia³.
}
\author{
Micha³ Danaj
}

