% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary.R
\name{prel_factors}
\alias{prel_factors}
\title{Podstawowe statystyki faktorów}
\usage{
prel_factors(dane)
}
\arguments{
\item{dane}{}
}
\value{

}
\description{
Wiele z faktorów zaczytanych bezpoœrednio z pliku nie nadaje siê do modelowania. Funkcja umo¿liwia wstêpn¹ 
analizê faktorów.
}
\author{
Micha³ Danaj
}

