% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary.R
\name{prel_logical}
\alias{prel_logical}
\title{Wyoewietla wektory logiczne}
\usage{
prel_logical(dane)
}
\arguments{
\item{dane}{}
}
\value{

}
\description{
Po wczytaniu danych z pliku tekstowego nie powinno byae wektorów logicznych w innych przypadkach ni?
pusty wektor (tego nie jestem tak na prawde teraz pewny). Funkcja wypisuje wszystkie wektory logiczne
?eby mo?na by^3o je obejrzeae.
}
\author{
Micha^3 Danaj
}
