% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary.r
\name{prel_numeric_one_value}
\alias{prel_numeric_one_value}
\title{Wyoewietla zmienne z wartooeci^1 o du?ym udziale w rozk^3adzie}
\usage{
prel_numeric_one_value(dane, rate = 0.98)
}
\arguments{
\item{dane}{Dane.}

\item{rate}{Graniczny udzia^3 jedenej wartooeci.}
}
\value{

}
\description{
Wyoewietla nazwy zmiennych, dla których istenieje taka wartooeae, ?e jej udzia^3 przekracza \code{rate}
próby.
}
\author{
Micha^3 Danaj
}
