% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary.R
\name{prel_numeric_sd0}
\alias{prel_numeric_sd0}
\title{Wyoewietla zmienne numeryczne z wariancj^1 == 0}
\usage{
prel_numeric_sd0(dane)
}
\arguments{
\item{dane}{}
}
\value{

}
\description{
Wyoewietla zmienne numeryczne z jedn^1 wartooeci^1.
}
\author{
Micha^3 Danaj
}
