% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_assign.R
\name{przypisz}
\alias{przypisz}
\title{Przypisuje wartoœci zdefiowane w \code{bucket}}
\usage{
przypisz(x, bucket, interpol = FALSE)
}
\arguments{
\item{x}{Wektor zmiennych numerycznych.}

\item{bucket}{\code{data.frame} z kolumnami:
\itemize{
        \item od. Dolny kraniec przedzia³u.
        \item do Górny kreniec przedzia³u.
        \item srodek Œrodek przedzia³u.
        \item fitted Wartoœæ przypisana do danego przedzia³u.
    }}

\item{interpol}{Czy przeprowadziæ interpolacjê?}
}
\value{
Wektor wartoœci.
}
\description{
Znajduje przedzia³ opisany w \code{bucket}, do którego nale¿y \code{x} i 
przypisuje wartoœæ z \code{bucket_fitted}. Ewentualnie wykonuje interpolacjê.
}
\details{
Jeœli \code{interpor=TRUE}, wykonuje interpolacjê, której wêz³ami s¹ 
œrodki przedzia³ów, pomiêdzy którymi le¿y \code{x}, a wartoœciami
\code{bucket$fitted} odpowiadaj¹ce tym œrodkom.
}
\examples{
		n<-1000;
		x1<-rnorm(n);
		y<-(x1+rnorm(n))<0;
		
		b<-buckety_br(x1,y,10);
		b$fitted<-b$br;
		
		y_fitted<-przypisz(x1, b, interpol=FALSE)
		plot(x1, y_fitted)
		
		y_fitted<-przypisz(x1, b, interpol=TRUE)
		plot(x1, y_fitted) 
}
\author{
Micha³ Danaj
}
\seealso{
\code{\link{przypisz2}}
}

