% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classing_assign.R
\name{przypisz}
\alias{przypisz}
\title{Przypisuje wartości zdefiowane w \code{bucket}}
\usage{
przypisz(x, bucket, interpol = FALSE, round = TRUE, digits = 15)
}
\arguments{
\item{x}{Wektor zmiennych numerycznych.}

\item{bucket}{\code{data.frame} z kolumnami:
\itemize{
        \item od. Dolny kraniec przedziału.
        \item do Górny kreniec przedziału.
        \item srodek Środek przedziału.
        \item fitted Wartość przypisana do danego przedziału.
    }}

\item{interpol}{Czy przeprowadzić interpolację?}

\item{round}{Czy zaokrąglać. Domyślnie tak.}

\item{digits}{Do której cyfry znaczącej zaokrąglać. Domyślnie do 15-tej.}
}
\value{
Wektor wartości.
}
\description{
Znajduje przedział opisany w \code{bucket}, do którego należy \code{x} i 
przypisuje wartość z \code{bucket_fitted}. Ewentualnie wykonuje interpolację.
}
\details{
Jeśli \code{interpor=TRUE}, wykonuje interpolację, której węzłami są 
środki przedziałów, pomiędzy którymi leży \code{x}, a wartościami
\code{bucket$fitted} odpowiadające tym środkom.

Aby zapobiec skutkom błedów numerycznych, domyślnie wartości \code{x}, jak i \code{bucket$srodek},
\code{bucket$od}, \code{bucket$do} zaokrąglone są do 15-tej cyfry znaczącej.
}
\examples{
		n<-1000;
		x1<-rnorm(n);
		y<-(x1+rnorm(n))<0;
		
		b<-buckety_br(x1,y,10);
		b$fitted<-b$br;
		
		y_fitted<-przypisz(x1, b, interpol=FALSE)
		plot(x1, y_fitted)
		
		y_fitted<-przypisz(x1, b, interpol=TRUE)
		plot(x1, y_fitted) 
}
\seealso{
\code{\link{przypisz}}

\code{\link{mapuj}}

\code{\link{przypisz2}}
}
\author{
Michał Danaj
}
