% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{przypisz_z_listy}
\alias{przypisz_z_listy}
\title{Na podstawie listy z bucketami przypisuje odpowiednie wartoœci}
\usage{
przypisz_z_listy(bucket_list, data, vars = names(bucket_list),
  colname = "fitted", varname_sufix = colname)
}
\arguments{
\item{bucket_list}{Lista bucketów.}

\item{data}{Dane.}

\item{vars}{Zmienne do przypisania.}

\item{colname}{Nazwa kolumny z bucketa, na podstawie którego nale¿y dopisaæ wartoœci.}

\item{varname_sufix}{Sufiks do nazw nowych zmiennych.}
}
\value{

}
\description{
Na podsatwie listy \code{bucket_list} przypisuje zmiennych \code{vars} z tabeli \code{data}
wartoœci znajduj¹ce siê w kolumnie \code{fitted} bucketów. Wartoœci przypisuje do kolumn
o nazwie z po³¹cznia nazw \code{vars} i \code{varname_sufix}.
}
\author{
Micha³ Danaj
}

