% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{reg_nieparam}
\alias{reg_nieparam}
\title{Rysuje lokalnie wyg³adzon¹ funckjê.
 
 Rysuje i zwraca statystyki dla bucketów.}
\usage{
reg_nieparam(score, default, buckets = 100, wytnij = 0, span = 0.7,
  degree = 2, plot = TRUE, plt_type = "br", new = TRUE,
  col_points = "black", col_line = "darkblue", index = FALSE,
  glm = FALSE, col_glm = "green", ...)
}
\arguments{
\item{score}{Wektor zmiennych numerycznych.}

\item{default}{Wektor zmiennej dwumianowej.}

\item{buckets}{Liczba bucketów, na ile nele¿y podzieliæ \code{score}.}

\item{wytnij}{Ile krañcowych obserwacji wyci¹æ.}

\item{span}{Wspó³czynnik wyg³adzania. Szegó³y w funkcji \code{\link[locfit]{locfit}}}

\item{degree}{Stopieñ wielomianu do lokalnego wyg³adzania. Szegó³y w funkcji \code{\link[locfit]{locfit}}}

\item{plot}{Czy rysowaæ wykres.}

\item{plt_type}{jeœli \code{br}, to na osi OY bêdzie BR. W przeciwnym razie bêdzie logit(BR)}

\item{new}{Czy rysowaæ wykres od nowa.}

\item{col_points}{Kolor punktów.}

\item{col_line}{Kolor lini.}

\item{index}{jeœli \code{TRUE}, na osi OX bêd¹ numery kolejnych bucketów.
W przeciwnym razie na osi OX bêd¹ wartoœci \code{score}.}

\item{glm}{czy rysowaæ dopasowanie modelu logistycznego do zmiennej.}

\item{col_glm}{kolor wykresu z modelu logistycznego.}

\item{...}{dodatkowe parametry.}
}
\description{
Rysuje lokalnie wyg³adzon¹ funckjê.
 
 Rysuje i zwraca statystyki dla bucketów.
}
\examples{
	n<-1000;
	x1<-rnorm(n);
	x2<-x1+rnorm(n);
	y<-(x1+rnorm(n))<0;
	
	reg_nieparam(x1,y, buckets=20)
	reg_nieparam(x2,y, buckets=20, new=FALSE, col_line="green",col_points="green")
}
\author{
Micha³ Danaj
}

