% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{step_bez_kor}
\alias{step_bez_kor}
\title{Wynik funkcji add1, po usuniêciu zmiennych skorelowanych}
\usage{
step_bez_kor(data, model, target_var_name = "target", cor_threshold = 0.75)
}
\arguments{
\item{data}{\code{data.frame} ze zmiennymi do dodania do modelu}

\item{model}{model do rozszerzenia}

\item{target_var_name}{nazwa zmiennej z targetem}

\item{cor_threshold}{graniczna wartoœæ korelacji, po przekroczeniu której zmienna jest uwuwana z analiz}
}
\value{
Statystyki
}
\description{
Sprawdza stopieñ korelacji zmiennych z danych \code{data} ze zmiennymi z modelu \code{model}. Jeœli korelacja
przekracza \code{cor_threshold}, to zmienna usuwana jest z dalszych analiz. Dla pozosta³ych zmiennych 
stosowana jest funkcja \code{\link{add1}}. W danych nie mo¿e byæ braków danych.
TODO Funkcja nie przetestowana, jeszcze do dopracowania!!!
}
\author{
Micha³ Danaj
}

