% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate_anal_stats}
\alias{univariate_anal_stats}
\title{Dyskretyzuje zmienną i wylicza dla niej statystyki.}
\usage{
univariate_anal_stats(x, y, czas, proby = rep(TRUE, length(y)),
  interactive = FALSE, min_bucket = floor(0.05 * length(x)),
  breaks = NULL, mapping = NULL, forceContinous = FALSE,
  special_val = numeric_var_treatment.params$special_val,
  NA_subst = numeric_var_treatment.params$NA_substit, span = 0.9)
}
\arguments{
\item{x}{zmienna, po której procedura będzie sortować.}

\item{y}{zmienna odpowiedzi.}

\item{czas}{podział na okresy czasowe.}

\item{proby}{\code{data.frame}, w którym każda kolumna jest wektorem logicznym,
zawierającym informacje, czy obserwacja należy do danej próby.}

\item{interactive}{TRUE, jeśli zmienna ma być dyskretyzowana interaktywnie. W
przeciwnym razie, co jest wartością domyślną, dyskretyzacja
jest automatyczna.}

\item{min_bucket}{minimalna liczba obserwacji w buckecie, przy dzieleniu drzewem.}

\item{breaks}{zamiast automatycznego dzielenia, można podać wartości przedziałów (from,to].}

\item{mapping}{zamiast automatycznego dzielenia, można podać mapowanie.}

\item{forceContinous}{wymusza potraktowanie zmiennej jako ciągłą, mimo że liczba
unikalnych wartości jest mniejsza niż \code{discrete_threshold}.}

\item{special_val}{Wartości specjalne do usunięcia z automatycznego podziału. Będą traktowane jako zmienne
kategoryczne.}

\item{NA_subst}{wartość jaka ma być przypisana w miejsce braków danych. Dalsze analizy
będą przeporwadzone w standardowy sposób. Jeśli jednak wartość \code{NA_subst} zostanie
    dodana do \code{special_val}, zostanie ona potraktowana jako wartość dyskretna.}

\item{span}{Parametr wygładzający funkcji \code{\link[locfit]{locfit}}.}
}
\description{
Wylicza statystyki i zwraca je w postaci listy.
}
\author{
Michał Danaj
}
