% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate_anal_stats2}
\alias{univariate_anal_stats2}
\title{Robi rozkład zmiennej po czasie (lub innym podziale)}
\usage{
univariate_anal_stats2(x_discr, y, czas, estim)
}
\arguments{
\item{x_discr}{zmienna objaśniająca.}

\item{y}{zmienna odpowiedzi.}

\item{czas}{Czas.}

\item{estim}{wartość wyestymowana przez model.}
}
\description{
Wylicza liczności dla każdego poziomu \code{x_discr} oraz średnią wartość \code{y} oraz \code{esitm}
w podziale na zadane grupy czasowe \code{czas} (lub podział innego typu).
}
