% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.r
\name{univariate_anal_stats3}
\alias{univariate_anal_stats3}
\title{Wylicza AR po czasie i w podziale na zadane próby}
\usage{
univariate_anal_stats3(score, y, czas, proby)
}
\arguments{
\item{score}{zmienna, po której procedura będzie sortować.}

\item{y}{zmienna odpowiedzi.}

\item{czas}{podział na okresy czasowe.}

\item{proby}{\code{data.frame}, w którym każda kolumna jest wektorem logicznym,
zawierającym informacje, czy obserwacja należy do danej próby.}
}
\description{
Wylicza AR po czasie i w podziale na zadane próby
}
