% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate_loop}
\alias{univariate_loop}
\title{Grupuje zmienne i generuje dla nich listę statystyk}
\usage{
univariate_loop(x_df, y, vsub_bool = rep(TRUE, ncol(x_df)),
  vsub_list = NULL, vsub_rola = NULL, proby = rep(TRUE, length(y)),
  czas = rep(1, length(y)), ...)
}
\arguments{
\item{x_df}{\code{data.frame} ze zmiennymi do anliz.}

\item{y}{\code{vector} z wartościami zmiennej odpowiedzi.}

\item{vsub_bool}{\code{vector} logiczny określający, które zmienne z \code{df} powinny zostać zanalizowane.}

\item{vsub_list}{\code{vector} z nazwami zmiennych z \code{df}, które powinny zostać zanalizowane.}

\item{vsub_rola}{\code{vector} z rolą każdej ze zmiennych. Zostaną zanalizowane zmienne, dla których rola 
ustawiona jest na \code{'explanatory'}.}

\item{proby}{\code{data.frame} z wektorami logicznymi, określającymi podpróby na których należy przeanalizować podpróby.}

\item{czas}{\code{vector} ze zmienną czasową, po której zostaną podzielone analizy.}

\item{...}{dodatkowe parametry do funkcji \code{\link{univariate_anal_stats}}.}
}
\value{
lista ze statstykami dla każdej ze zmiennych określonych przez któryś z wektorów vsub\*. Nazwy elementów listy
	są nazwami zmiennych.
}
\description{
Funkcja grupuje zmienne ciągłe, zgodnie z domyślnymi parametrami. Na grupach oraz na zmiennych kategorycznych 
wylicza statytystyki i przedstawia je w formie listy, osobno dla każdej ze zmiennych. Wykorzystuje do tego funkcję
\code{\link{univariate_anal_stats}}.
}
\author{
Michal Danaj
}
\seealso{
\code{\link{univariate_anal_stats}}
}

