% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate_stats_new_data}
\alias{univariate_stats_new_data}
\title{Wylicza statystyki dla zmiennej score'owej i objaœnianej wg zadanych bucketów}
\usage{
univariate_stats_new_data(buckets, x, y, czas, proby = rep(TRUE, length(y)))
}
\arguments{
\item{buckets}{\code{data.frame} z podzia³em zmiennej na buckety.}

\item{x}{zmienna, po której procedura bêdzie sortowaæ.}

\item{y}{zmienna odpowiedzi.}

\item{czas}{podzia³ na okresy czasowe.}

\item{proby}{\code{data.frame}, w którym ka¿da kolumna jest wektorem logicznym,
zawieraj¹cym informacje, czy obserwacja nale¿y do danej próby.}
}
\description{
Wylicza statystyki i zwraca je w postaci listy. Do wyliczenia \code{AR} u¿ywa jako zmiennej
score'owej oryginalnych wartoœci BR przypisanych do bucketu. Tzn, mimo ¿e na nowych danych kolejnoœæ
bucketów, sortuj¹c je po br mo¿e byæ inna ni¿ oryginalnie, to stosowana jest oryginalna kolejnoœæ.
}
\author{
Micha³ Danaj
}

