% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDBinom.r
\name{usun_konce}
\alias{usun_konce}
\title{Usuwa krañcowe wartoœci.}
\usage{
usun_konce(score, prob = 0.01, weights = NULL)
}
\arguments{
\item{score}{Wektor wartoœci numerycznych.}

\item{prob}{Jak¹ czêœæ obserwacj nale¿y usun¹æ z jednego krañca.}

\item{weights}{Wagi.}
}
\value{
Zwraca wektor z indeksami elementów, które powinny zostaæ usuniête.
}
\description{
Zwraca indeksy \code{prob} najmniejszych i \code{prob} najwiêkszych wartoœci. Jeœli
jednak nie jest mo¿liwe ustalenie co najwy¿ej \code{prob} obserwacji, nie 
zostaje wybrana ¿adna. Ma to zapobiec sytuacji, gdy np. najmniejsz¹ wartoœci¹ jest 0
i w próbie znajduje siê 10\% takich obserwacji, a my chcemy usun¹æ tylko 1\%.
}
\examples{
x<-sort(rnorm(10));
x
usun<-usun_konce(x, prob=0.01);
x[-usun]

#usuwa tylko obserwacjê z prawego krañca
x2<-c(rep(min(x),5),x[5:10])
x2
usun<-usun_konce(x2, prob=0.01);
x2[-usun]
}
\author{
Micha³ Danaj
}

